% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar-filter.R
\name{ar2eps}
\alias{ar2eps}
\title{nested function
can we export Rcpp function with roxgen2?
convert AR(1) process into its eps with given rho}
\usage{
ar2eps(Yt, rho)
}
\description{
nested function
can we export Rcpp function with roxgen2?
convert AR(1) process into its eps with given rho
}
