% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_acf.R
\name{ar_acf}
\alias{ar_acf}
\title{ACF of the Autoregressive Model}
\usage{
ar_acf(alpha, rho, dT, N, nlim = 100)
}
\arguments{
\item{rho}{parameter of MA model}
}
\value{
vector of length N
}
\description{
ACF of the Autoregressive Model
}
\details{
Yt = (1-rho) Yt-1 + rho Xt, where Xt is fBM process with parameter alpha
}
