% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cochran_mq.R
\name{cochran_mq}
\alias{cochran_mq}
\title{Cochran's multivariate Q-test for heterogeneity.}
\usage{
cochran_mq(est, ve)
}
\arguments{
\item{est}{A \verb{k x p} matrix of independent estimates of a common parameter.}

\item{ve}{A \verb{(p x p x k)}-dimensional array of corresponding variance estimates.}
}
\value{
A list with elements:
\describe{
\item{\code{Q}}{The value of the test statistic.}
\item{\code{df}}{The degrees of freedom of the test.}
\item{\code{pval}}{The p-value of the test.}
\item{\code{coef}}{The combined estimate of the parameter under \code{H0}.}
\item{\code{vcov}}{The variance estimate of the combined estimator.}
}
}
\description{
Suppose we have \code{k} independent estimators of \code{p}-dimensional vectors \verb{theta_1, ..., theta_k}.  Under the null hypothesis \code{H0} that the \code{theta_i} are all the same, the Cochran Q-statistic is the precision-weighted sum-of-squares of the estimators.  Asymptotically it follows a chi-square distribution with \code{p*(k-1)} degrees of freedom.
}
