% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cochranMQ.R
\name{cochranMQ}
\alias{cochranMQ}
\title{Cochran's multivariate Q-test for heterogeneity.}
\usage{
cochranMQ(est, ve)
}
\arguments{
\item{est}{A \code{k x p} matrix of independent estimates of a common parameter.}

\item{ve}{A \code{(p x p x k)}-dimensional array of corresponding variance estimates.}
}
\value{
A list with elements:
\describe{
  \item{\code{Q}}{The value of the test statistic.}
  \item{\code{df}}{The degrees of freedom of the test.}
  \item{\code{pval}}{The p-value of the test.}
  \item{\code{est}}{The combined estimate of the parameter under \code{H0}.}
  \item{\code{ve}}{The variance estimate of the combined estimator.}
}
}
\description{
Under \code{H0}, \code{k} independent estimators are of a common \code{p}-diemensional parameter; their precision-weighted sum-of-squares follows asymptotically a \code{chi^2} distribution with \code{k-p} degrees of freedom.
}
