% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cochranQ.R
\name{cochranQ}
\alias{cochranQ}
\title{Cochran's Q-test for heterogeneity.}
\usage{
cochranQ(est, se)
}
\arguments{
\item{est}{A vector of independent estimates of a common parameter.}

\item{se}{A corresponding vector of standard errors.}
}
\value{
The value of the test statistic, its p-value, and the weighted mean.
}
\description{
Under \code{H0}, \code{k} independent estimators are of a common parameter; their precision-weighted sum-of-squares follows asymptotically a chi^2 distribution with \code{k-1} degrees of freedom.
}
