% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite-prof.R
\name{composite.prof}
\alias{composite.prof}
\title{Profile Likelihood for Composite}
\usage{
composite.prof(suff, Y, X, acf, ds, noSigma = FALSE)
}
\arguments{
\item{Y}{Position of process}

\item{X}{Position of mean}

\item{acf}{Acf of downsampled increment process}

\item{ds}{Downsampling rate}

\item{noSigma}{flag for Sigma matrix}
}
\value{
a list with the following elements:
\itemize{
   \item \code{Betahat = sum(t(Xi)V^{-1}Xi)^{-1}sum(t(Xi)V^{-1}Yi})
   \item \code{S = sum(t(Yi-Xi*Betahat)V^{-1}(Yi-Xi*Betahat))}
   \item \code{ldV = log(|V|)}
   \item \code{n.ds = nrow(Y.ds) - 1}
   \item \code{loglik}
}
}
\note{
We assume that the weights [w1, w2, ..., wK] are equal, thus it is omitted in estimation
}
