% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite-suff.R
\name{composite.suff}
\alias{composite.suff}
\title{nested function for sufficient statistics of composite likelihood
relationship between the nrow of Yt and acf: length(acf) = floor(Yt/ds) - 1}
\usage{
composite.suff(Y, X, acf, ds)
}
\description{
nested function for sufficient statistics of composite likelihood
relationship between the nrow of Yt and acf: length(acf) = floor(Yt/ds) - 1
}
