% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csi_resid.R
\name{csi_resid}
\alias{csi_resid}
\title{Calculate the residuals of a CSI process.}
\usage{
csi_resid(dX, drift, acf, Sigma)
}
\arguments{
\item{dX}{Matrix of trajectory increments.}

\item{drift}{Matrix of the same size as \code{dX} of increment drift in each coordinate per timepoint (see 'Details').}

\item{acf}{Autocorrelation vector of length \code{nrow(dX)} (see 'Details').}

\item{Sigma}{Covariance matrix of size \verb{ncol(dX) x ncol(dX)}.}
}
\value{
A residual matrix of the same size as \code{dX}.
}
\description{
Calculate the residuals of a CSI process.
}
\details{
The residuals are calculated as

\if{html}{\out{<div class="sourceCode">}}\preformatted{Z = toeplitz(acf)^\{-1/2\} \%*\% (dX - drift) \%*\% Sigma^\{-1/2\},
}\if{html}{\out{</div>}}

where the "square roots" correspond to the Cholesky decomposition for the row-dependence matrix \code{toeplitz(acf)}, and the eigen decomposition for the column-dependence matrix \code{Sigma}.
}
\examples{
# simulate data from the fbm model
alpha <- .8

dt <- 1/60
N <- 1800
ndim <- 2

Xt <- csi_sim(drift = matrix(0, N-1, ndim),
              acf = fbm_acf(alpha, dt, N-1),
              Sigma = diag(ndim),
              X0 = rep(0, ndim))


# compute the residuals
Z <- csi_resid(dX = apply(Xt, 2, diff),
               drift = matrix(0, N-1, ndim),
               acf = fbm_acf(alpha, dt, N-1),
               Sigma = diag(ndim))
}
