% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csi_sim.R
\name{csi_sim}
\alias{csi_sim}
\title{Simulate increments of a CSI process.}
\usage{
csi_sim(drift, acf, Sigma, X0, nsim = 1, fft = TRUE, nkeep, tol = 1e-06)
}
\arguments{
\item{drift}{Matrix of \verb{(nobs-1) x ndim} increment drift terms (see 'Details').}

\item{acf}{Increment autocorrelation vector of length \code{nobs-1} (see 'Details').}

\item{Sigma}{Variance matrix of size \verb{ndim x ndim}.}

\item{X0}{Initial position vector of length \code{ndim}.}

\item{nsim}{Number of trajectories to simulate.}

\item{fft, nkeep, tol}{Optional arguments to \code{\link[SuperGauss:rnormtz]{SuperGauss::rnormtz()}}.}
}
\value{
The simulated trajectories.  An array of size \verb{nobs x ndim x nsim} or a matrix of size \verb{nobs x ndim} when \code{nsim = 1}.
}
\description{
Simulate increments of a CSI process.
}
\details{
Given an \verb{(nobs-1) x ndim} matrix of iid standard normals \code{Z}, an \verb{nobs x ndim} trajectory matrix is generated by

\if{html}{\out{<div class="sourceCode">}}\preformatted{dX = toeplitz(acf)^\{1/2\} \%*\% Z \%*\% Sigma^\{1/2\} + drift
X = apply(rbind(X0, dX), 2, cumsum)
}\if{html}{\out{</div>}}
}
