% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample-naive.R
\name{downsample.naive}
\alias{downsample.naive}
\title{fBM inference using composite downsampling}
\usage{
downsample.naive(Xt, dT, Tz, ds, pos = 1, varCal = FALSE, se = FALSE,
  trans = FALSE)
}
\arguments{
\item{Xt}{Time series for inference}

\item{dT}{Interobservation time}

\item{Tz}{Toeplitz-object of size floor(nrow(Xt)/ds)-1}

\item{ds}{Downsampling rate}

\item{pos}{position of downsampling, integer between 1 and ds, default to be 1.}

\item{varCal}{logic, return the covariance matrix of MLE estimate if true}

\item{se}{logic, return the standard error of MLE instead of covariance matix if true}

\item{trans}{logic, return the 'normalized' parameters}
}
\description{
We assume that time series dY follows MN(Beta * dT, V_incr.fBM, Sigma),
thus downsampled dY_ds follows MN(Beta * ds * dT, V_ds.incr.fBM, Sigma).
}
