% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample}
\alias{downsample}
\title{Downsample the observation of time series.}
\usage{
downsample(Xt, r, position = 1)
}
\arguments{
\item{Xt}{One or two-column matrix of observations.}

\item{r}{Integer of downsampling rate.}

\item{position}{Integer of position index for downsampled series, smaller than or equal to `r`.}
}
\value{
An matrix of downsampled observation.
}
\description{
Downsample the observation of time series.
}
\details{
Sample observations from original time series \eqn{X = [X_0, X_2, X_3, ..., X_N]} at constant rate `r`. For given position index `i` finally returns \eqn{X_{ds} = [X_i, X_{i+r}, X_{i+2r}, ...]}.
}
\examples{
Xt <- apply(matrix(rnorm(2000), 2, 1000), 2, cumsum) # generates a 2-d series of Brownian motion.
Xt_ds2 <- downsmaple(Xt, r = 2)

}
