% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_functions.R
\name{drift_functions}
\alias{drift_functions}
\alias{drift_linear}
\alias{drift_none}
\alias{drift_quadratic}
\title{Predefined drift functions for \code{csi_model} objects.}
\usage{
drift_linear(phi, dt, N)

drift_none(phi, dt, N)

drift_quadratic(phi, dt, N)
}
\arguments{
\item{phi}{Vector of kernel parameters in the original basis.}

\item{dt}{Interobservation time \eqn{\Delta t} = 1/fps (positive scalar).}

\item{N}{Number of observations (positive integer).}
}
\description{
Predefined drift functions for \code{csi_model} objects.
}
\details{
Models inheriting from the \code{csi_model} base class must specify a drift function having argument signature \verb{function(phi, dt, N)} and returning an \verb{N x n_drift} matrix of drift basis functions for the trajectory \emph{increments} (as opposed to the original positions).  The predefined drift functions provided here can account for linear drift, quadratic drift, or no drift at all.
}
