% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyn-err.R
\name{dyn.infer}
\alias{dyn.infer}
\title{inference for dynamic error model with fBM process}
\usage{
dyn.infer(Yt, dt, acf, ConfInt = FALSE)
}
\arguments{
\item{Yt}{Observation of trajectory}

\item{dt}{interobservation}

\item{acf}{Toeplitz space}

\item{ConfInt}{logical, default to be FALSE}
}
\value{
if \code{ConfInt} = \code{TRUE} return list containing estimation and length of error bar, 
return estimation only otherwise.
}
\details{
static error model:
Y(t) = int_0^tau [G(t-s) * beta + Z(t-s) * Sigma^{1/2}] ds
increment form:
dY(t) = dG(t) * beta + X(t, tau) * Sigma^{1/2}
where X(t, tau) is the integral of increment of fBM process
}
\note{
requires package \code{LMN} and \code{SuperGauss}
}
