% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/far_acf.R
\name{far_acf}
\alias{far_acf}
\title{ACF of the Autoregressive Model with fBM Noise}
\usage{
far_acf(alpha, rho, dT, N, nlim = 100)
}
\arguments{
\item{alpha}{TBD}

\item{rho}{Parameter of AR model}

\item{dT}{TBD}

\item{N}{TBD}
}
\value{
Vector of length N
}
\description{
ACF of the Autoregressive Model with fBM Noise
}
\details{
\code{Yt = (1-rho) Yt-1 + rho Xt}, where \code{Xt} is fBM process with parameter \code{alpha}.
}
