% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/far_fit.R
\name{far_fit}
\alias{far_fit}
\title{Fit the fractional AR(1) model.}
\usage{
far_fit(dX, dT, Tz, var_calc = TRUE, theta0, ...)
}
\arguments{
\item{dX}{One or two-column matrix of increments.}

\item{dT}{Interobservation time.}

\item{Tz}{Optional Toeplitz matrix for intermediate calculations.}

\item{var_calc}{If \code{TRUE}, also estimate variance matrix.}

\item{theta0}{Length-2 vector of initial values for \code{(alpha, rho)}.  Default value is \code{(1, 0)}.}

\item{...}{Additional arguments to \code{\link[stats]{optim}}.}
}
\value{
Vector of coefficients and possibly variance matrix on the transformed scale (see Details).
}
\description{
Fit the fractional AR(1) model.
}
\details{
The fractional AR(1) model has the form
\deqn{
\Delta X_n = (1-\rho) \Delta Z_n + \rho \Delta X_{n-1},
}
where \eqn{\Delta Z_n} are increments of a 1D or 2D fBM process. The MLE and variance estimate are calculated on the transformed scale defined by \code{trans(rho) = logit(1-rho/2)}, \code{trans(mu) = mu}, \code{\link{trans_alpha}}, and \code{\link{trans_Sigma}}.
}
