% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farma_acf.R
\name{farma_acf}
\alias{farma_acf}
\title{Calculate the ACF for the farma model.}
\usage{
farma_acf(alpha, phi = numeric(), rho = numeric(), dT, N, m = 30)
}
\arguments{
\item{alpha}{Power law exponent of the fBM process. A scalar between 0 and 2.}

\item{phi}{A vector of \code{p >= 0} autoregressive (AR) coefficients.}

\item{rho}{A vector of \code{q >= 0} moving-average (MA) coefficients (see 'Details').}

\item{dT}{Interobservation time \eqn{\Delta t} = 1/fps (positive scalar).}

\item{N}{Number of observations (positive integer).}

\item{m}{Number of MA coefficients used for approximating the ARMA filter (see \strong{Details}).}
}
\value{
A vector of \code{N} autocorrelation values.
}
\description{
Compute the autocovariance of increment process of a fARMA(p,q) model (see \strong{Details}).
}
\details{
The fARMA(p,q) model is of following form:
\deqn{
Y_n = \sum_{i=1}^p \phi_i Y_{n-i} + \sum_{j=0}^q \rho_j X_{n-j}
}{
Y[n] = \phi_1 Y[n-1] + ... + \phi_p Y[n-p] + \rho_0 X[n] + ... + \rho_q X[n-q]
}
where residuals \eqn{X[n]} follow the fBM process with parameter \code{alpha} (see \code{\link[=fbm_acf]{fbm_acf()}}), and \code{rho_0 = 1 - sum(phi) - sum(rho)}.

This function returns the autocovariance of stationary increment process \eqn{\Delta Y[n] = Y[n] - Y[n-1]}.  The \code{ARMA(p,q)} filter is approximated with \code{m} moving-average terms using the method of Brockwell & Davis (1991).
}
\examples{
# ACF parameters
alpha <- 0.8
dT <- 1/60
N <- 1800
phi <- 0.1
rho <- c(0.1, 0.05)
# ACF of farma model
farma_acf(alpha, phi, rho, dT, N)
}
\references{
Ling, Y., Lysy, M., Seim, I., Newby, J.M., Hill, D.B., Cribb, J., and Forest, M.G. "Measurement error correction in particle tracking microrheology." (2019). \url{https://arxiv.org/abs/1911.06451}.
}
