% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farma_model.R
\name{farma_model}
\alias{farma_model}
\title{Constructor for farma model object.}
\description{
Constructor for farma model object.
}
\details{
Let \code{X_n} denote the true position of an fBM process at time \verb{t = n dt}.  The farma(p,q) error model describes the measured position \code{Y_n} at time \verb{t = n dt} as

\if{html}{\out{<div class="sourceCode">}}\preformatted{Y_n = sum_\{i=1\}^p phi_i Y_\{n-i\} + sum_\{j=0\}^q rho_j X_\{n-j\},
}\if{html}{\out{</div>}}

with \verb{rho_0 = 1 - sum_\{i=1\}^p phi_i - sum_\{j=1\}^q rho_j}.  The resulting MSD as a function of \code{phi} and \code{rho} is what gets passed to the \link{csi_model} base class to construct the \code{farma_model} derived class.
}
\examples{
# simulate data from a farma(1,1) model
alpha <- .8
phi <- .1
rho <- .1

dt <- 1/60
N <- 1800
ndim <- 2

Xt <- csi_sim(drift = matrix(0, N-1, ndim),
              acf = farma_acf(alpha, phi = phi, rho = rho, dt, N-1),
              Sigma = diag(ndim),
              X0 = rep(0, ndim))


# create farma(1,1) model object
model <- farma_model$new(Xt = Xt, dt = dt, order = c(1, 1), drift = "linear")

# evaluate loglikelihood
model$loglik(phi = c(alpha = alpha, phi = phi, rho = rho),
             mu = rep(0, ndim),
             Sigma = diag(ndim))

}
\section{Super class}{
\code{\link[subdiff:csi_model]{subdiff::csi_model}} -> \code{farma_model}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{phi_names}}{Kernel parameter names.  The character vector \verb{(alpha, phi_1, ..., phi_p, rho_1, ..., rho_q)}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-farma_model-acf}{\code{farma_model$acf()}}
\item \href{#method-farma_model-trans}{\code{farma_model$trans()}}
\item \href{#method-farma_model-itrans}{\code{farma_model$itrans()}}
\item \href{#method-farma_model-get_subdiff}{\code{farma_model$get_subdiff()}}
\item \href{#method-farma_model-new}{\code{farma_model$new()}}
\item \href{#method-farma_model-clone}{\code{farma_model$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="drift"><a href='../../subdiff/html/csi_model.html#method-csi_model-drift'><code>subdiff::csi_model$drift()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="fisher"><a href='../../subdiff/html/csi_model.html#method-csi_model-fisher'><code>subdiff::csi_model$fisher()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="fit"><a href='../../subdiff/html/csi_model.html#method-csi_model-fit'><code>subdiff::csi_model$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="get_omega"><a href='../../subdiff/html/csi_model.html#method-csi_model-get_omega'><code>subdiff::csi_model$get_omega()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="get_vcov"><a href='../../subdiff/html/csi_model.html#method-csi_model-get_vcov'><code>subdiff::csi_model$get_vcov()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="itrans_full"><a href='../../subdiff/html/csi_model.html#method-csi_model-itrans_full'><code>subdiff::csi_model$itrans_full()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="loglik"><a href='../../subdiff/html/csi_model.html#method-csi_model-loglik'><code>subdiff::csi_model$loglik()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="msd"><a href='../../subdiff/html/csi_model.html#method-csi_model-msd'><code>subdiff::csi_model$msd()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="nlp"><a href='../../subdiff/html/csi_model.html#method-csi_model-nlp'><code>subdiff::csi_model$nlp()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="nu_hat"><a href='../../subdiff/html/csi_model.html#method-csi_model-nu_hat'><code>subdiff::csi_model$nu_hat()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="resid"><a href='../../subdiff/html/csi_model.html#method-csi_model-resid'><code>subdiff::csi_model$resid()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="sim"><a href='../../subdiff/html/csi_model.html#method-csi_model-sim'><code>subdiff::csi_model$sim()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="trans_full"><a href='../../subdiff/html/csi_model.html#method-csi_model-trans_full'><code>subdiff::csi_model$trans_full()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-farma_model-acf"></a>}}
\if{latex}{\out{\hypertarget{method-farma_model-acf}{}}}
\subsection{Method \code{acf()}}{
Increment autocorrelation function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{farma_model$acf(phi, dt, N)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phi, dt, N}}{See \link{csi_model}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-farma_model-trans"></a>}}
\if{latex}{\out{\hypertarget{method-farma_model-trans}{}}}
\subsection{Method \code{trans()}}{
Transform kernel parameters from regular to computational basis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{farma_model$trans(phi)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phi}}{See \link{csi_model}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-farma_model-itrans"></a>}}
\if{latex}{\out{\hypertarget{method-farma_model-itrans}{}}}
\subsection{Method \code{itrans()}}{
Transform kernel parameters from computational to regular basis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{farma_model$itrans(psi)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{psi}}{See \link{csi_model}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-farma_model-get_subdiff"></a>}}
\if{latex}{\out{\hypertarget{method-farma_model-get_subdiff}{}}}
\subsection{Method \code{get_subdiff()}}{
Transform parameters from computational basis to subdiffusion parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{farma_model$get_subdiff(omega)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{omega}}{See \link{csi_model}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector with named elements \code{alpha} and \code{logD}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-farma_model-new"></a>}}
\if{latex}{\out{\hypertarget{method-farma_model-new}{}}}
\subsection{Method \code{new()}}{
Class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{farma_model$new(Xt, dt, order = c(0, 0), m = 50, drift = "linear", n_drift)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Xt, dt, drift, n_drift}}{See \link{csi_model}.}

\item{\code{order}}{Vector of two nonnegative integers specifying the number of autoregressive and moving-average terms, respectively.}

\item{\code{m}}{Order of the moving-average approximation (see \code{\link[=farma_acf]{farma_acf()}}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-farma_model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-farma_model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{farma_model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
