% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbm-acf.R
\name{fbm.acf}
\alias{fbm.acf}
\title{Acf function for the increment of fBM process.}
\usage{
fbm.acf(alpha, dT, N)
}
\arguments{
\item{alpha}{fBM parameter, range from 0 to 2.}

\item{dT}{interobservation time.}

\item{N}{number of samples.}
}
\value{
vector of length N.
}
