% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbm_acf.R
\name{fbm_acf}
\alias{fbm_acf}
\title{Calculate the autocorrelation function for the fBM model.}
\usage{
fbm_acf(alpha, dT, N)
}
\arguments{
\item{alpha}{Power law exponent.  A scalar between 0 and 2.}

\item{dT}{Interobservation time \eqn{\Delta t =} 1/fps (positive scalar).}

\item{N}{Length of observations (positive integer).}
}
\value{
A vector of \code{N} autocorrelation values.
}
\description{
Compute the autocorrelation of fBM increments of equally-spaced observations (see \strong{Details}).
}
\details{
The autocorrelation of fBM increments at lag \eqn{h} is given by
\deqn{
\textrm{acf}(h) = \Delta t^\alpha/2 \times \left(|n+1|^\alpha + |n-1|^\alpha - 2|n|^\alpha\right).
}{
acf(n) = \Delta t^\alpha/2 * (|n+1|^\alpha + |n-1|^\alpha - 2|n|^\alpha). 
}
}
