% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbm_msd.R
\name{fbm_msd}
\alias{fbm_msd}
\title{Calculate the mean square displacement of the fBM model.}
\usage{
fbm_msd(t, alpha)
}
\arguments{
\item{t}{Vector of timepoints at which to calculate the MSD.}

\item{alpha}{Subdiffusion exponent. A scalar between 0 and 2.}
}
\value{
Vector of MSD values the same length as \code{t}.
}
\description{
Calculate the mean square displacement of the fBM model.
}
\details{
If \code{X_t} is an fBM process, then\preformatted{msd_X(t) = t^alpha.
}
}
