% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdl_fit.R
\name{fdl_fit}
\alias{fdl_fit}
\title{Fit the fractional dynamic localization error model.}
\usage{
fdl_fit(dX, dT, tau, sigma2, Tz, var_calc = TRUE, penalty = TRUE, theta0,
  ...)
}
\arguments{
\item{dX}{one or two-column matrix of trajectory increments.}

\item{dT}{Interobservation time.}

\item{tau}{Plug-in coefficient, requires estimation if missing,}

\item{sigma2}{Plug-in coefficient, requires estimation if missing,}

\item{Tz}{Optional Toeplitz matrix for intermediate calculations.}

\item{var_calc}{If \code{TRUE}, also estimate variance matrix.}

\item{penalty}{logic, employ a small penalty on \code{(tau, sigma2)} when \code{TRUE}}

\item{theta0}{Length-3 vector of initial values for \code{(alpha, tau, sigma)}.  Default value is \code{(1, .1, .1)}.}

\item{...}{Additional arguments to \code{\link[stats]{optim}}.}
}
\value{
Vector of coefficients and possibly variance matrix on the transformed scale (see Details).
}
\description{
Fit the fractional dynamic localization error model.
}
\details{
The fractional AR(1) model has the form
\deqn{
X_n = 1/\tau \int_0^\tau Z_(n+s) ds + \sigma e_{n},
}
where \eqn{Z_n} is a 1D or 2D fBM process. The MLE and variance estimate are calculated on the transformed
scale defined by \code{trans(tau) = logit(tau)}, \code{trans(sigma) = log(sigma)}, \code{trans(mu) = mu}, \code{\link{trans_alpha}}, and
\code{\link{trans_Sigma}}.
When put \code{tau} = 0, this model becomes fractional localization error model.
When put \code{sigma2} = 0, this model becomes fractional dynamic error model.
When put \code{(tau, sigma2)} = 0, this model becomes fractional Brownian model.
}
