% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fds_fit.R
\name{fds_fit}
\alias{fds_fit}
\title{Fit fBM model with composite-likelihood downsampling.}
\usage{
fds_fit(dX, dT, ds, method = c("naive", "comp"), Tz, var_calc = TRUE)
}
\arguments{
\item{dX}{One or two-column matrix of increments.}

\item{dT}{Interobservation time.}

\item{Tz}{Optional Toeplitz matrix for intermediate calculations.}

\item{var_calc}{If \code{TRUE}, also estimate variance matrix.}

\item{type}{type of downsampling method, "naive" means using basic downsampling method, "comp" means using composite likelihood (see Details).}

\item{...}{Additional arguments to \code{\link[stats]{optim}}.}
}
\value{
Vector of coefficients and possibly variance matrix on the transformed scale (see Details).
}
\description{
Fit fBM model with composite-likelihood downsampling.
}
\details{
We assume that time series dX follows MN(Beta * dT, V_fBM(alpha, dT), Sigma), thus downsampled dY_ds follows MN(Beta * ds * dT, V_fBM(alpha, dT*ds), Sigma).
}
