% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdyn_acf.R
\name{fdyn_acf}
\alias{fdyn_acf}
\title{ACF2 of fBM + Dynamic Error}
\usage{
fdyn_acf(alpha, tau, dT, N)
}
\arguments{
\item{alpha}{Subdiffusion exponent}

\item{tau}{Ratio of time-window over \code{dT}, range from 0 to 1.}

\item{dT}{Vector of time points \eqn{ \{\Delta t, 2\Delta t, ..., N\Delta t\} }}

\item{N}{length}
}
\description{
ACF2 of fBM + Dynamic Error
}
\details{
this function returns the MSD of \eqn{Y_t}, the integral of fBM process \eqn{X_t} with subdiffusion 
exponent \eqn{\alpha} \deqn{Y_t = \int_{0}^{\tau} X(t-s)ds}. The expression of the MSD is
\deqn{\frac{(t+\tau)^\alpha + (t-\tau)^\alpha - 2t^\alpha - 2\tau^\alpha}{(\alpha+1)(\alpha+2)}}
}
\examples{
fdyn.msd(alpha = 0.8, tau = 1/600, t = (1:200) * 1/60)
}
