% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floc_acf.R
\name{floc_acf}
\alias{floc_acf}
\title{Calculate the ACF for the fLOC model.}
\usage{
floc_acf(alpha, tau, sigma2, dT, N)
}
\arguments{
\item{alpha}{Power law exponent of fBM process. A scalar between 0 and 2.}

\item{tau}{The ratio between camera shutter open time and interobservation time \code{dT} (see \strong{Details}). A scalar between 0 and 1.}

\item{sigma2}{The Magnitude of static errors (see \strong{Details}). A positive scalar.}

\item{dT}{Interobservation time \eqn{\Delta t =} 1/fps (positive scalar).}

\item{N}{Length of observations (positive integer).}
}
\value{
A vector of \code{N} autocorrelation values.
}
\description{
Compute the autocovariance for the increments of Savin & Doyle's localization model with fBM process (see \strong{Details}).
}
\details{
The Savin & Doyle's localization error model with fBM process (floc model) is of following form:
\deqn{
Y_n = 1/\tau \int_0^\tau X_{n  \Delta t+s} ds + \sigma e_{n}
}{
Y[n] = 1/\tau integral_0^\tau X_{n  \Delta t+s} ds + \sigma e[n]
}
where \eqn{X_n} is an fBM process. 

The integral part measures the average displacement of particle trajectories during \eqn{[n \Delta t, n\Delta t + \tau]} and this term is called dynamic error.
\eqn{e[n]} is a standard white noise, and this term is called static error.

This function returns the autocovariance function of increments of \eqn{Y[n]}.
}
\examples{
# ACF parameters
alpha <- 0.8
dT <- 1/60
N <- 1800
tau <- 0.1
sigma2 <- 0.001
# ACF of floc model
floc_acf(alpha, tau, sigma2, dT, N)
}
\references{
Savin, Thierry, and Patrick S. Doyle. "Static and dynamic errors in particle tracking microrheology." Biophysical journal 88.1 (2005): 623-638.
}
