% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floc_fit.R
\name{floc_fit}
\alias{floc_fit}
\title{Fit the fBM model with Savin & Doyle's Localization Error.}
\usage{
floc_fit(dX, dT, Tz, var_calc = TRUE)
}
\arguments{
\item{dX}{One or two-column matrix of increments. Increments of Gaussian CSI processes.}

\item{dT}{Interobservation time \eqn{\Delta t =} 1/fps (positive scalar).}

\item{Tz}{Optional Toeplitz matrix for intermediate calculations.}

\item{var_calc}{If \code{TRUE}, also estimate variance matrix.}
}
\value{
Vector of coefficients and possibly variance matrix on the transformed scale (see Details).
}
\description{
Fit the fBM model with Savin & Doyle's Localization Error.
}
\details{
The fBM + localization error (fLOC) model has the form
\deqn{
Y_n = 1/\tau \int_0^\tau X_{n+s} ds + \sigma e_{n},
},
where \eqn{X_n} is a 1D or 2D fBM process
\deqn{
X_t = \mu t + \Sigma^{1/2} Z_t,
}.
where \eqn{Z_t} consists of \code{q = 1,2} iid fBM processes.
When put \code{tau} = 0, this model becomes fractional localization error model.
When put \code{sigma2} = 0, this model becomes fractional dynamic error model.
When put \code{(tau, sigma2)} = 0, this model becomes fractional Brownian model.
}
\examples{
fits <- floc_fit(dX, dT, Tz, var_calc = TRUE)

# generating time series
dT <- 1/60
N <- 1800
acf1 <- fbm_acf(0.8, dT, N)
dX <- rSnorm(n = 2, acf = acf1)
Tz <- Toeplitz(n = N)
}
