% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fma_acf.R
\name{fma_acf}
\alias{fma_acf}
\title{ACF of the Moving Average Model with fBM Noise}
\usage{
fma_acf(alpha, rho, dT, N)
}
\arguments{
\item{alpha}{TBD}

\item{rho}{parameter of MA model}

\item{dT}{TBD}

\item{N}{TBD}
}
\value{
Vector of length N
}
\description{
ACF of the Moving Average Model with fBM Noise
}
\details{
\code{Yt = (1-rho) Xt + rho Xt-1}, where \code{Xt} is fBM process with parameter \code{alpha}.
}
