% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fma_fit.R
\name{fma_fit}
\alias{fma_fit}
\alias{fma_trunc}
\title{Fit the fractional MA(1) model.}
\usage{
fma_fit(dX, dT, nlag, Tz, var_calc = TRUE, ...)

fma_trunc(dX, dT, rho)
}
\arguments{
\item{dX}{One or two-column matrix of increments.}

\item{dT}{Interobservation time.}

\item{nlag}{Number of lags (see Details).}

\item{Tz}{Optional Toeplitz matrix for intermediate calculations.}

\item{var_calc}{If \code{TRUE}, also estimate variance matrix.}

\item{...}{Additional arguments to \code{\link[stats]{optim}}.}
}
\value{
Vector of coefficients and possibly variance matrix on the transformed scale (see Details).
}
\description{
Fit the fractional MA(1) model.
}
\details{
The fractional MA(1) model has the form
\deqn{
\Delta X_n = (1-\rho) \Delta Z_n + \rho \Delta Z_{n-1},
}
where \eqn{\Delta Z_n} are increments of a 1D or 2D fBM process. The MLE and variance estimate are calculated on the transformed scale defined by \code{trans(rho) = logit(1-rho/2)}, \code{trans(mu) = mu}, \code{\link{trans_alpha}}, and \code{\link{trans_Sigma}}.
}
