% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd_acf.R
\name{fsd_acf}
\alias{fsd_acf}
\title{Calculate the autocorrelation of the fSD model.}
\usage{
fsd_acf(alpha, tau, sigma2, dt, N)
}
\arguments{
\item{alpha}{Subdiffusion exponent of the underlying fBM process. A scalar between 0 and 2.}

\item{tau}{The ratio between camera shutter open time and interobservation time \code{dt} (see 'Details'). A scalar between 0 and 1.}

\item{sigma2}{The magnitude of the static error (see 'Details'). A positive scalar.}

\item{dt}{Interobservation time \eqn{\Delta t} = 1/fps (positive scalar).}

\item{N}{Number of observations (positive integer).}
}
\value{
A vector of \code{N} autocorrelation values.
}
\description{
Compute the autocorrelation of the Savin & Doyle (2005) localization error model with fBM increments (see 'Details').
}
\details{
Let \code{X_t} denote the position of an fBM process at time \code{t}.  The Savin-Doyle localization error model describes the observed position \code{Y_n} at time \code{t = n * dt} as\preformatted{Y_n = sigma * eps_n + 1/tau * int_0^tau X_\{n*dt + s\} ds,
}

where \verb{eps_n ~iid N(0,1)} is a Gaussian white noise process.

This function returns the autocorrelation of the stationary process \verb{dY_n = Y_\{n+1\} - Y_n}.
}
\examples{
# ACF parameters
alpha <- 0.8
dt <- 1/60
N <- 1800
tau <- 0.1
sigma2 <- 0.001
# ACF of fsd model
fsd_acf(alpha, tau, sigma2, dt, N)
}
\references{
Savin, T., and Doyle, P.S. "Static and dynamic errors in particle tracking microrheology." Biophysical Journal 88.1 (2005): 623-638.
}
