% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd_fit.R
\name{fsd_fit}
\alias{fsd_fit}
\title{Fit the fSD model.}
\usage{
fsd_fit(dX, dt, drift = c("linear", "none", "quadratic"), vcov = TRUE)
}
\arguments{
\item{dX}{One or two-column matrix of increments. Increments of Gaussian CSI processes.}

\item{dt}{Interobservation time \eqn{\Delta t} = 1/fps (positive scalar).}

\item{drift}{Character string specifying one of the \emph{preset} drift types "linear", "none", or "quadratic".  Custom drift functions are not supported through this simplified interface.  See \link{csi_model} for details.}

\item{var_calc}{Logical; if \code{TRUE}, also estimate the variance matrix.}
}
\value{
A vector of estimated parameters on transformed scale (See \code{\link[=fsd_model]{fsd_model()}}). If \code{vcov}, a list with components:
\describe{
\item{coef}{A vector of estimated parameters on transformed scale.}
\item{vcov}{A matrix of estimated covariance of parameters on transformed scale.}
}
}
\description{
Fit Savin & Doyle's localization error model with fBM process (See 'Details').
}
\details{
To avoid issues with the boundary of the parameter support (on the regular scale), the optimization is conducted with a penalty term\preformatted{penalty(psi) = log(1+exp(psi2)) + log(1+exp(-psi2)) - psi3.
}

The Savin & Doyle's localization error model with fBM process (fsd model) has the following form:
\deqn{
Y_n = 1/\tau \int_0^\tau X_{n \Delta t +s} ds + \sigma e_{n}
}{
Y[n] = 1/\tau integral_0^\tau X_{n  \Delta t +s} ds + \sigma e[n]
}
where \eqn{X_n} is an \verb{q = 1,2} dimensional fBM model (See \code{\link[=fbm_fit]{fbm_fit()}}).

The integral part measures the average displacement of particle trajectories during \eqn{[n \Delta t, n\Delta t + \tau]} and this term is called dynamic error.
\eqn{e_n} is a standard white noise, and this term is called static error.

When put \code{tau} = 0, this model becomes fractional static error model.
When put \code{sigma2} = 0, this model becomes fractional dynamic error model.
When put \verb{(tau, sigma2)} = 0, this model becomes fractional Brownian model.

Optimization is done by \code{\link[=optimize]{optimize()}}. It works better when parameters are re-parametrized into unrestricted form (See \code{\link[=fsd_model]{fsd_model()}}).
}
\examples{
# generating 2-d stationary time series
alpha <- 0.8
dt <- 1/60
N <- 1800
acf1 <- fbm_acf(alpha, dt, N)
dX <- rSnorm(2, acf = acf1)
# Toeplitz object for internal computation
Tz <- Toeplitz(n = N)
# Fit the fsd model
fsd_fit(dX, dt, Tz, vcov = TRUE)
}
\references{
Savin, Thierry, and Patrick S. Doyle. "Static and dynamic errors in particle tracking microrheology." Biophysical Journal 88.1 (2005): 623-638.
}
