% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd_model.R
\name{fsd_model}
\alias{fsd_model}
\title{Constructor of fsd model object.}
\usage{
fsd_model()
}
\value{
An list of class \code{csi_class} with components:
\describe{
\item{\code{theta_names}}{A vector of parameter names.}
\item{\code{acf}}{An ACF function with three arguments: vector of parameters; interobservation time \eqn{\Delta t}; length \code{N}.}
\item{\code{theta_trans}}{Transformation function that converts original ACF parameter into unrestricted parameters.}
\item{\code{theta_itrans}}{Inverse Transformation function that converts unrestricted parameters into original ACF parameter.}
\item{\code{penalty}}{Penalty function for unrestricted parameters.}
}
}
\description{
Constructor of fsd model object.
}
\details{
Constructor function that generates a list of fsd model for purpose of fitting.
}
\examples{
# generate a fsd model object
model <- fsd_model()
# we can modify the elements of generated object
model$theta_names[1] <- "gamma"
}
