% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd_msd.R
\name{fsd_msd}
\alias{fsd_msd}
\title{Calculate the mean square displacement of the fSD model.}
\usage{
fsd_msd(t, dt, alpha, tau, sigma2)
}
\arguments{
\item{t}{Vector of timepoints at which to calculate the MSD.}

\item{dt}{Interobservation time \eqn{\Delta t} = 1/fps (positive scalar).}

\item{alpha}{Subdiffusion exponent of the underlying fBM process. A scalar between 0 and 2.}

\item{tau}{The ratio between camera shutter open time and the interobservation time \code{dt}. A scalar between 0 and 1.}

\item{sigma2}{The magnitude of the static error. A positive scalar.}
}
\value{
Vector of MSD values the same length as \code{t}.
}
\description{
Compute the MSD of the Savin & Doyle (2005) localization error model with fBM increments (see \code{\link[=fsd_acf]{fsd_acf()}}).
}
\details{
Let \code{X_t} denote the true position of an fBM process at time \code{t}.  The Savin-Doyle localization error model describes the measured position \code{Y_n} at time \verb{t = n dt} as\preformatted{Y_n = sigma * eps_n + 1/(tau*dt) * int_0^(tau*dt) X_\{n dt - s\} ds,
}

where \verb{eps_n ~iid N(0,1)} is a Gaussian white noise process.  The resulting MSD as a function of \code{tau} and \code{sigma2 = sigma^2} is what gets passed to the \link{csi_model} base class to construct the \code{fsd_model} derived class.
}
\references{
Savin, T., and Doyle, P.S. "Static and dynamic errors in particle tracking microrheology." Biophysical Journal 88.1 (2005): 623-638.
}
