% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbe.R
\docType{data}
\name{hbe}
\alias{hbe}
\title{Simulated HBE dataset.}
\format{
A dataset containing 76 simulated 2D particle trajectories sampled at 60 Hz.  The four columns of the data are:
\describe{
\item{\code{id}}{Particle ID (1-76).}
\item{\code{t}}{Time at which position was recorded (in seconds).}
\item{\code{x}}{Particle position on \code{x}-axis at time \code{t} (in microns).}
\item{\code{y}}{Particle position on \code{y}-axis at time \code{t} (in microns).}
}
}
\source{
Hill, D. B., Vasquez, P. A., Mellnik, J., McKinley, S. A., Vose, A., Mu, F., Henderson, A. G., Donaldson, S. H., Alexis, N. E., Boucher, R. C., and Forest, M. G. (2014), "A Biophysical Basis for Mucus Solids Concentration as a Candidate Biomarker for Airways Disease", PLoS ONE(9), e87681.
}
\usage{
hbe
}
\description{
Simulated HBE dataset.
}
\keyword{datasets}
