% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessEst.R
\name{hess.estimator}
\alias{hess.estimator}
\title{Fisher Information Estimation for MLE variance}
\usage{
hess.estimator(func, x, ...)
}
\arguments{
\item{func}{log-likelihood function}

\item{x}{estimation of parameter}
}
\value{
vector of length of x
}
\description{
MLE estimation asymptotically follows normal distribution with mean unbiased and 
variance Fisher Information, which equals the negative inverse of Hessian matrix under 
most circumstance.
}
\note{
require package \code{numDeriv}
}
