% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transFunc.R
\name{itransFunc}
\alias{itransFunc}
\title{Inverse function of transFunc}
\usage{
itransFunc(theta, trans)
}
\arguments{
\item{theta}{transformed parameters}

\item{trans}{logic, true means \code{theta} is transformed}
}
\value{
list containing {alpha, Beta, Sigma}
}
