% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\alias{ilogit}
\title{Generalized logistic and inverse-logistic transformations.}
\usage{
logit(x, min = 0, max = 1)

ilogit(x, min = 0, max = 1)
}
\arguments{
\item{x}{A numeric scalar or vector.}

\item{min, max}{Numeric scalar or vectors denoting the bounds of the transformations.}
}
\value{
The logit or inverse-logit transoformation.
}
\description{
Generalized logistic and inverse-logistic transformations.
}
\details{
The generalized logistic transformation is
\preformatted{
logit(x) = log(p/(1-p)),    p = (x-min)/(max - min),
}
and its inverse is
\preformatted{
ilogit(x) = min + (max-min)/(1+exp(-x).
}
}
