% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsc_resid.R
\name{lsc_resid}
\alias{lsc_resid}
\title{Residuals of a location-scale stationary increments model.}
\usage{
lsc_resid(dX, dT, mu, acf, Sigma)
}
\arguments{
\item{dX}{Matrix of increments where each column is a time series.}

\item{dT}{Interobservation time.}

\item{mu}{Vector of linear drift in each direction.}

\item{acf}{Autocorrelation, a vector of length \code{nrow(dX)}.}

\item{Sigma}{Covariance matrix of size \code{ncol(dX) x ncol(dX)}.}
}
\value{
A residual matrix of the same size as \code{dX}.
}
\description{
Residuals of a location-scale stationary increments model.
}
\details{
The residuals are calculated as
\deqn{
Z = Toeplitz(acf)^{-1/2} (dX - mu dT) Sigma^{-1/2}
}
where the "square roots" correspond to the Cholesky decomposition for the Toeplitz row-dependence matrix, and the eigen decomposition for the column-dependence matrix.
}
