% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma-filter.R
\name{ma.filter}
\alias{ma.filter}
\title{MA(1) Filter}
\usage{
ma.filter(dX, dT, Tz, varCal = FALSE, se = FALSE, trans = FALSE)
}
\arguments{
\item{dX}{Time series for inference}

\item{dT}{Interobservation time}

\item{Tz}{Toeplitz-object of size N}

\item{varCal}{logic, return the covariance matrix of MLE estimate if true}

\item{se}{logic, return the standard error of MLE instead of covariance matix if true}

\item{trans}{logic, return the 'normalized' parameters}
}
\description{
Model: dY(t) = (1-rho) dX(t) + rho dX(t-1)
}
