% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma3_fit.R
\name{ma3_fit}
\alias{ma3_fit}
\title{Fit the MA(3) model.}
\usage{
ma3_fit(dX, dT, var_calc = TRUE, ...)
}
\arguments{
\item{dX}{one or two-column matrix of trajectory increments.}

\item{dT}{Interobservation time.}

\item{var_calc}{If \code{TRUE}, also estimate variance matrix.}

\item{...}{Additional \code{control} arguments to \code{stats::optim}.}
}
\value{
Vector of coefficients and possibly variance matrix on the transformed scale (see Details).
}
\description{
Fit the MA(3) model.
}
\details{
The fractional MA(3) model has the form
\deqn{
\Delta X_n =  \mu \Delta t + \Sigma^{1/2} Z_n,
}
\deqn{
Z_n = \epsilon_n + \phi_1 \epsilon_{n-1} + \phi_2 \epsilon_{n-2} + \phi_3 \epsilon_{n-3},
}
where \eqn{\epsilon_n \sim N(0,1)}. The MLE and variance estimate are calculated on the transformed scale defined by \code{trans(phi) = phi}, \code{trans(mu) = mu}, and \code{\link{trans_Sigma}}.  Fitting is done by assuming that \code{eps[0] = eps[-1] = eps[-2] = 0}, leading to a simple formula for the model residuals via the \code{stats::filter} function.  Preliminary parameter estimates obtained from \code{stats::arima} are passed on to \code{optim}.
}
