% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma_acf.R
\name{ma_acf}
\alias{ma_acf}
\title{Calculate the autocorrelation of an \code{MA(q)} filter applied to a stationary process.}
\usage{
ma_acf(acf, rho = numeric())
}
\arguments{
\item{acf}{Vector of length \code{N + q} of autocorrelations of the original stationary process.}

\item{rho}{Vector of \code{q >= 0} coefficients of the \code{MA(q)} filter (see 'Details').}
}
\value{
A vector of length \code{N} giving the autocorrelation of the filtered process.
}
\description{
Calculate the autocorrelation of an \code{MA(q)} filter applied to a stationary process.
}
\details{
Let \code{X[n]} denote the observation of the original process at time \code{n}, and \code{Y[n]} denote the corresponding observation of the filtered process.  The \code{MA(q)} filter model is defined as\preformatted{Y[n] = X[n] + sum_\{j=1\}^q rho[j] X[n-j].
}
}
