% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-cov.R
\name{mle.cov}
\alias{mle.cov}
\title{Covariance Matrix for MLE estimator}
\usage{
mle.cov(loglik.func, theta, ...)
}
\arguments{
\item{loglik.func}{likelihood function}

\item{theta}{MLE estimation}
}
\description{
compute the covariance matrix of MLE estimator
}
