% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModePoly.R
\name{modePoly}
\alias{modePoly}
\title{Wrapper function for Mode polynomial}
\usage{
modePoly(lambda, nSteps, tol)
}
\arguments{
\item{nSteps}{number of steps}

\item{tol}{tolerence}
}
\value{
vector of length of \code{lambda}
}
\description{
MODEPOLY given a polynomial with distinct roots a_0 < ... < a_N, find its
local modes using the golden section method programmed in c++.
}
