% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msd.R
\name{msd}
\alias{msd}
\title{Mean Square Displacement}
\usage{
msd(Xt, dT, nlag = 30, msd.only = TRUE)
}
\arguments{
\item{Xt}{trajectory}

\item{dT}{interobservation time, required when computing slope \code{alpha}}

\item{nlag}{order of MSD of interested, default 20}

\item{msd.only}{logical, return both computed msd and slope of msd when TRUE}
}
\value{
vector of length \code{nlag}
}
