% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msd_fit.R
\name{msd_fit}
\alias{msd_fit}
\title{Sample estimate of mean squared displacement.}
\usage{
msd_fit(Xt, nlag, demean = TRUE, dT = NULL, msd)
}
\arguments{
\item{Xt}{Vector or matrix of trajectories (each one is a column).}

\item{nlag}{Number of MSD lags to calculate.}

\item{demean}{Logical; whether or not to remove the mean of \code{Xt} as estimated by linear drift.}

\item{dT}{Optional interobservation time.  If supplied, estimates subdiffusion parameters \code{(alpha, D)} by regressing MSD over time on log-log scale.}

\item{msd}{Optionally pre-computed sample MSD, in which case it is used for regression estimate of \code{(alpha, D)}.}
}
\value{
Sample MSD vector of length \code{nlag}, or list with elements \code{msd}, \code{alpha}, \code{D}.
}
\description{
Sample estimate of mean squared displacement.
}
