% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_pval.R
\name{norm_pval}
\alias{norm_pval}
\alias{ad_pval}
\alias{sw_pval}
\alias{bc_pval}
\title{P-value calculation for various normality tests.}
\usage{
ad_pval(x, stdn = TRUE)

sw_pval(x)

bc_pval(x)
}
\arguments{
\item{x}{Vector of observations.}

\item{stdn}{Logical, whether or not to test against standard normal distribution, or take mean and variance to be unknown.}
}
\value{
p-value of the normality test.
}
\description{
P-value calculation for various normality tests.
}
\details{
The following normality tests are currently implemented:
\describe{
  \item{\code{ad_pval}}{The Anderson-Darling test, which assumes iid observations.  This is merely a wrapper to \code{\link[nortest]{ad.test}}.}
  \item{\code{sw_pval}}{A modified Shapiro-Wilk test, which also assumes iid observations.  For sample sizes less than 5000, this function is merely a wrapper to \code{\link[stats]{shapiro.test}}.  For sample sizes greater than 5000, the algorithm due to Royston (1992) used in \code{\link[stats]{shapiro.test}} breaks down.  The modification proposed here is to (i) randomly divide the sample into \code{K} roughly equal parts each less than 5000, (ii) calculate the Shapiro-Wilk p-value for each, and (iii) note that the minimum of \code{K} independent p-values has a \code{Beta(1, K)} distribution.  It is this second-stage p-value whic is returned by \code{sw_pval}.}
  \item{\code{bc_pval}}{The so-called "Berkowitz' Correlation" test.  That is, An autoregressive model of the form
\preformatted{
x[n] = rho * x[n-1] + z[n],    z[n] ~iid N(0,1)
}
is fit to the data, and a likelihood ratio test against \code{H0: rho = 0} is performed.}
}
}
\note{
The tests above **only** check for normality, not that \code{x} is iid \code{N(0,1)}.  Such modifications will need to be made before the package is released.
}
\references{
Royston, P. "Approximating the Shapiro-Wilk W-test for non-normality." \emph{Statistics and Computing} 2:3 (1992): 117-119. \url{https://doi.org/10.1007/BF01891203}.

Berkowitz, J. "Testing density forecasts, with applications to risk management." \emph{Journal of Business and Economic Statistics} 19:4 (2001): 465-474. \url{https://doi.org/10.1198/07350010152596718}.
}
