% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prony-coeff.R
\name{prony.coeff}
\alias{prony.coeff}
\title{Coefficients of position-process representation of the Prony-GLE.}
\usage{
prony.coeff(lambda, r, C, nsteps = 100, tol = 0)
}
\arguments{
\item{lambda}{Vector of inverse decorrelation times of the GLE force term.}

\item{r}{Optional vector of pre-computed OU mean-reversion parameters.  Avoids calling the mode-finding routine.}

\item{C}{Optional vector of pre-computed OU scale coefficients.  Can't pass \code{C} without \code{r}.  When both \code{r} and \code{C} are supplied the function does nothign.  This is mainly for the convenience of calling \code{prony.coeff(...)} from other functions.}

\item{nsteps}{Number of steps in mode-finding golden search algorithm.}

\item{tol}{Relative tolerance in mode-finding golden search algorithm.  For numerical stability, default is 0 such that \code{nsteps} are always used.}
}
\value{
A list with elements \code{r} and \code{C}, containing the \code{length(lambda)-1} and \code{length(lambda)} vectors of mean-reversion parameters and scale factors for the BM + OU representation of the Prony-GLE.
}
\description{
Coefficients of position-process representation of the Prony-GLE.
}
