% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prony_msd.R
\name{prony_msd}
\alias{prony_msd}
\title{MSD of the Prony-GLE model.}
\usage{
prony_msd(t, lambda, ...)
}
\arguments{
\item{t}{times at which to calculate the MSD.}

\item{lambda}{coefficients of the sum of OU processes.}

\item{...}{Additional arguments to \code{\link[=prony_coeff]{prony_coeff()}} for obtaining the BM-OU coefficients.}
}
\value{
Vector of mean square displacements.
}
\description{
MSD of the Prony-GLE model.
}
\details{
The Prony-GLE model satisfies the integro-differential equation
\deqn{
F_t - \int_{-\infty}^t \gamma(t-s) \dot X_s ds = 0
}
\deqn{
acf_F(t) = k_B T \gamma(t),
}
where \eqn{T} is temperature, \eqn{k_B} is Boltzmann's constant, and the memory kernel is a sum of exponentials:
\deqn{
\gamma(t) = 1/K sum_{n=1}^K exp(-\lambda_n * t).
}
The solution process is of the form
\deqn{
X_t = C_0 B_t + \sum_{i=1}^{K-1} C_i W_{it},
}
where \eqn{d W_{it} = -\rho_i W_{it} dt + d B_{it}} are Ornstein-Uhlenbeck processes all independent of each other and of \eqn{B_t}.
}
\note{
Since temperature is not provided, the result is only proportional to the desired GLE, such that PUT IN CORRECT EQUATION.
}
