% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rouse-acf.R
\name{rouse.acf}
\alias{rouse.acf}
\title{Rouse ACF}
\usage{
rouse.acf(alpha, tau, K, N, dt, ...)
}
\arguments{
\item{alpha}{why use 1/alpha in this function?}
}
\description{
autocorrelation for the Rouse anomalous diffusion model. This corresponds to a GLE with 0 potential and 0 mass with prony series memory kernel.
}
\details{
gamma(t) = sum_{n=1}^K 1/K * exp(-alpha_n * t), alpha_n = (n/K)^rho/tau.
the output is only proportional to the autocorrelation, in the sense that
acf_x(t) = (2 * k_B * T * K) * acf.
}
