% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rouse-msd.R
\name{rouse.msd}
\alias{rouse.msd}
\title{Rouse MSD}
\usage{
rouse.msd(t, rho, tau, K, ...)
}
\value{
vector
}
\description{
mean squared-displacement for the Rouse anomalous diffusion model.
}
\details{
this corresponds to a GLE with 0 potential and 0 mass with prony series memory kernel 
gamma(t) = sum_{n=1}^K 1/K * exp(-alpha_n * t), alpha_n = (n/K)^rho/tau.
the output is only proportional to the autocorrelation, in the sense that
msd_x(t) = (2 * k_B * T * K) * msd
}
