% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rouse_acf.R
\name{rouse_acf}
\alias{rouse_acf}
\title{ACF of the Rouse-GLE model.}
\usage{
rouse_acf(alpha, tau, K, dt, N, ...)
}
\arguments{
\item{alpha}{Subdiffusion coefficient between 0 and 1.}

\item{tau}{Shortest timescale of memory kernel.}

\item{K}{Number of relaxation modes.}

\item{dt}{Time between obserations.}

\item{N}{Number of observations.}

\item{...}{Additional arguments to pass to \code{\link[=prony_acf]{prony_acf()}}.}
}
\value{
Vector of autocorrelations.
}
\description{
Autocorrelation of the Rouse-GLE transient subdiffusion model.
}
\details{
The Rouse-GLE satisfies the integro-differential equation
\deqn{
F_t - \int_{-\infty}^t \gamma(t-s) \dot X_s ds = 0,
}
where \eqn{acf_F(t) = k_B T \gamma(t)} and
\deqn{
\gamma(t) = 1/K sum_{k=1}^K exp(-\lambda_k * t), \lambda_k = (k/K)^(1/alpha)/tau.
}

As the temperature \eqn{T} is not supplied, the output is only proportional to the autocorrelation, in the sense that \code{ACF_X(t) = k_B * T * rouse_acf}.
}
