% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rouse_msd.R
\name{rouse_msd}
\alias{rouse_msd}
\title{MSD of the Rouse-GLE model.}
\usage{
rouse_msd(t, alpha, tau, K, ...)
}
\arguments{
\item{t}{Vector of timepoints at which to calculate the MSD.}

\item{alpha}{Subdiffusion coefficient between 0 and 1.}

\item{tau}{Shortest timescale of memory kernel.}

\item{K}{Number of relaxation modes.}

\item{...}{Additional arguments to pass to \code{prony.msd}.}
}
\value{
Vector of mean square displacements.
}
\description{
Mean Squared Displacement of the Rouse-GLE transient subdiffusion model.
}
\details{
The Rouse-GLE satisfies the integro-differential equation
\deqn{
F_t - \int_{-\infty}^t \gamma(t-s) \dot X_s ds = 0,
}
where \eqn{acf_F(t) = k_B T \gamma(t)} and
\deqn{
\gamma(t) = 1/K sum_{n=1}^K exp(-\lambda_n * t), \lambda_n = (n/K)^(1/alpha)/tau.
}

As the temperature \eqn{T} is not supplied, the output is only proportional to the MSD, in the sense that \eqn{MSD_x(t) = (2 * k_B * T * K) * msd}.
}
