% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rouse-sub.R
\name{rouseSub}
\alias{rouseSub}
\title{rouse subdiffusive}
\usage{
rouseSub(rho, tau, K, tRange, alphaRouse = FALSE)
}
\arguments{
\item{tau}{relaxation timescale: min and max relaxation times of force F_t are tau * [1, K^(1/alpha)]}

\item{K}{number of modes in relaxation spectrum}

\item{tRange}{how to estimate the subdiffusive timescale.}

\item{alphaRouse}{logical, whether or not to use alpha = 1/rho or the least-squares estimate. Defaults to false.}

\item{alpha}{theoretical subdiffusion as K goes to infinity}
}
\value{
list containing tScale(the subdiffusion timescale), alpha(the subdiffusion exponent) and 
D(the diffusivity coefficient)
}
\description{
this function determines the timescale and coefficients of the subdiffusion
approximation for the Rouse-GLE model.
}
\details{
The Rouse model is a zero-mass, zero-potential GLE with force memory
kernel

acf_F(t) = 1/K sum( exp(- t * (1:K/K)^rho / tau) ).

As K gets large, it exhibits has transient subdiffusion:

MSD(t) ~ D * t^alpha, t \in tScale
         t,           t \notin tScale,
          
where alpha = 1/rho is the subdiffusion exponent and D is the diffusivity
constant.
}
\note{
Either 'force' or 'inflection' method.  When tRange is empty or
missing, it is the former, which uses tScale = tau * [1, K^(1/alpha)].  When tRange 
is a vector of length 2, the inflection
method uses the above property of transient subdiffusion, which means
that on the log-log scale, the slope of the MSD goes from 1 to alpha
< 1 to 1. Therefore, its acceleration starts at zero, becomes
negative, then positive, then returns to zero. The two inflection
points are thus the min and max of d2 log(MSD(t)) / d [log(t)]^2.
}
