% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwichEst.R
\name{sand.estimator}
\alias{sand.estimator}
\title{Sanwich Estimator}
\usage{
sand.estimator(func, x, ...)
}
\arguments{
\item{func}{log-likelihood function}

\item{x}{estimation of parameter}
}
\value{
vector of length of x
}
\description{
For incorrect likelihood function, its variance is canot be estimated using Hessian matrix, thus we 
use sanwich estimator to approximate the error bar.
}
\note{
require package \code{numDeriv}
}
