% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich-cov.R
\name{sandwich.cov}
\alias{sandwich.cov}
\title{Covariance Matrix of Sanwich Estimator}
\usage{
sandwich.cov(loglik.func, theta, ...)
}
\arguments{
\item{loglik.func}{list of component of log-likelihood function}

\item{theta}{estimation of parameter}
}
\description{
For incorrect likelihood function, its variance is canot be estimated using Hessian matrix,
thus we should use sanwich estimator to approximate the error bar.
}
