% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sand-inf.R
\name{sandwich.infer.prof}
\alias{sandwich.infer.prof}
\title{nested function for sandwich.infer
composite likelihood}
\usage{
sandwich.infer.prof(alpha, Yt, dt, acf, rate)
}
\description{
nested function for sandwich.infer
composite likelihood
}
