% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sand-inf.R
\name{sandwich.infer}
\alias{sandwich.infer}
\title{Downsampling Inference using Composite Likelihood and Sandwich Estimator}
\usage{
sandwich.infer(Yt, dt, acf, rate = 1, ConfInt = FALSE)
}
\arguments{
\item{Yt}{Observation of trajectory}

\item{dt}{interobservation}

\item{acf}{Toeplitz space}

\item{rate}{dowmsapling rate, default to be 1, meaning no downsampling}

\item{ConfInt}{logical, default to be FALSE}
}
\value{
if \code{ConfInt} = \code{TRUE} return list containing estimation and length of error bar, 
return estimation only otherwise.
}
\note{
requires package \code{LMN} and \code{SuperGauss}
}
