% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich-se.R
\name{sandwich.se}
\alias{sandwich.se}
\title{Sanwich Estimator}
\usage{
sandwich.se(func, theta, ...)
}
\arguments{
\item{func}{log-likelihood function}

\item{x}{estimation of parameter}
}
\value{
vector of length of x
}
\description{
For incorrect likelihood function, its variance is canot be estimated using Hessian matrix, 
thus we use sanwich estimator to approximate the error bar.
}
\note{
require package \code{numDeriv}
}
