% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sDiffEst.R
\name{sDiffEst}
\alias{sDiffEst}
\title{sDiffEst}
\usage{
sDiffEst(tSeq, msd, alpha, wgt)
}
\description{
this function calculates subdiffusion power law 
MSD(t) = D t^alpha
 from a discretized MSD least-squares estimate.
}
\details{
[alpha, D] = sDiffEst(tSeq, msd, wgt) calculates these coefficients
 from the regression
 
 'log(msd) ~  log(D) + alpha tSeq + wgt^(-1/2) eps.'
 
 When tSeq is of length 2 it denotes a range.
 alpha if provided is fixed.
 wgt if not provided is ones.
}
