% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-err.R
\name{stat.infer}
\alias{stat.infer}
\title{inference for static error model with fBM process}
\usage{
stat.infer(Yt, dt, acf, ConfInt = FALSE)
}
\arguments{
\item{Yt}{Observation of trajectory}

\item{dt}{interobservation}

\item{acf}{Toeplitz space}

\item{ConfInt}{logical, default to be FALSE}
}
\value{
if \code{ConfInt} = \code{TRUE} return list containing estimation and length of error bar, 
return estimation only otherwise.
}
\details{
static error model:
Y(t) = G(t) * beta + (Z(t) + a * I(t)) * Sigma^{1/2}
increment form:
dY(t) = dG(t) * beta + (dZ(t) + a * dI(t)) * Sigma^{1/2}
}
\note{
requires package \code{LMN} and \code{SuperGauss}
}
