% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subdiff-package.R
\docType{package}
\name{subdiff-package}
\alias{subdiff}
\alias{subdiff-package}
\title{Inference for Subdiffusive Particle Tracking.}
\description{
Tools for implementing various models for particle subdiffusion in biological fluids.
}
\author{
\strong{Maintainer}: Martin Lysy \email{mlysy@uwaterloo.ca}

Authors:
\itemize{
  \item Yun Ling
}

}
