% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid-functions.R
\name{subdiff-resid}
\alias{subdiff-resid}
\alias{fbm_resid}
\alias{fma_resid}
\alias{far_resid}
\alias{fds_resid}
\alias{fdl_resid}
\title{Model-based residual calculations.}
\usage{
fbm_resid(theta, dX, dT)

fma_resid(theta, dX, dT)

far_resid(theta, dX, dT)

fds_resid(theta, dX, dT, type, ds)

fdl_resid(theta, dX, dT, type)
}
\arguments{
\item{theta}{Vector of parameter values on the transformed scale returned by the corresponding \code{model_fit} procedure (see Details).}

\item{dX}{One or two-column matrix of increments.}

\item{dT}{Interobservation time.}

\item{type, ds}{Additional arguments specifying for the downsampling estimator (see \code{\link{fds_fit}}).}
}
\value{
A one or two-column matrix of white noise residuals (see Details).
}
\description{
Extracts the white-noise residuals from the Gaussian models provided by the package.
}
\details{
Each of the models provided by the package generates a normal data matrix \eqn{\Delta X} via the linear transformation
\deqn{
\Delta X = A Z B + C,
}
where the variance matrices \eqn{V = AA'} and \eqn{\Sigma = B'B} and mean matrix \eqn{C} are defined by the model parameters, and \eqn{Z} is a matrix of iid \eqn{N(0,1)} white-noise residuals.  The specific square-roots of the variance matrices are the Cholesky decomposition of \eqn{V} (time-domain variance), and the eigendecomposition of \eqn{\Sigma} (spatial-domain variance).
}
