% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_window.R
\name{time_window}
\alias{time_window}
\title{Search for time window in sample mean square displacement curves.}
\usage{
time_window(msd, tseq, error = 0.05)
}
\arguments{
\item{msd}{Vector or matrix of sample MSDs, each column corresponding to a different trajectory.}

\item{tseq}{Vector of time points at which the MSDs are recorded (see Details).}

\item{error}{Relative allowance for difference between msd and linear fit}
}
\value{
A 3-row matrix of \code{tmin}, \code{alpha} and \code{D} values
}
\description{
Search for time window in sample mean square displacement curves.
}
\details{
Time window is defined as the longest successive time period whose linear fitted msd is within a certain tolerence.
Time window is determined by two parameters \code{tmin} and \code{tmax}. 
In current version we simply assume that \code{tmax} is always out-of-observation and equals the end of experimental time scale, and focus on the search of \code{tmin}.
}
