% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_alpha.R
\name{trans_alpha}
\alias{trans_alpha}
\alias{itrans_alpha}
\title{Normalizing transformation for the subdiffusion coefficient.}
\usage{
trans_alpha(alpha)

itrans_alpha(gamma)
}
\arguments{
\item{alpha}{Subdiffusion coefficient on regular scale.}

\item{gamma}{Subdiffusion coefficient on normalized scale.}
}
\value{
Subdiffusion coefficient on the regular or normalized scale.
}
\description{
Normalizing transformation for the subdiffusion coefficient.
}
\details{
The normalizing transformation for the subdiffusion coefficient is
\deqn{
\gamma = logit(\alpha/2).
}
}
