% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_gamma.R
\name{trans_gamma}
\alias{trans_gamma}
\alias{itrans_gamma}
\title{Normalizing transformation for the subdiffusion coefficient.}
\usage{
trans_gamma(gamma)

itrans_gamma(igamma)
}
\arguments{
\item{gamma}{Subdiffusion coefficient on regular scale.}

\item{igamma}{Subdiffusion coefficient on normalized scale.}
}
\value{
Subdiffusion coefficient on the regular or normalized scale.
}
\description{
Normalizing transformation for the subdiffusion coefficient.
}
\details{
The normalizing transformation for the subdiffusion coefficient is
\deqn{
\theta = logit((\gamma+1)/2).
}
}
