% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_rho.R
\name{trans_rho}
\alias{trans_rho}
\alias{itrans_rho}
\title{Normalizing transformation for the AR/MA filters.}
\usage{
trans_rho(rho)

itrans_rho(eta)
}
\arguments{
\item{rho}{Filter parameter on regular scale.}

\item{eta}{Filter parameter on the normalized scale.}
}
\value{
Filter parameter on the regular or normalized scale.
}
\description{
Normalizing transformation for the AR/MA filters.
}
\details{
The normalizing transformation for the filter parameter is
\deqn{
\eta = logit((\rho+1)/2).
}
}
