% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_tau.R
\name{trans_tau}
\alias{trans_tau}
\alias{itrans_tau}
\title{Normalizing transformation for the dynamic error model.}
\usage{
trans_tau(tau)

itrans_tau(tau)
}
\arguments{
\item{tau}{Camera exposure time divided by the interobservation time.}

\item{tau}{Camera exposure parameter on the normalized scale.}
}
\value{
Camera exposure parameter on the regular or normalized scale.
}
\description{
Normalizing transformation for the dynamic error model.
}
\details{
The normalizing transformation for the camera exposure parameter is
\deqn{
\tau = logit(\tau).
}
}
