% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transFunc.R
\name{transFunc}
\alias{transFunc}
\title{transform parameters into standard form}
\usage{
transFunc(alpha, Beta, Sigma, trans)
}
\arguments{
\item{alpha}{fBM parameter}

\item{Beta}{linear drift parameter}

\item{Sigma}{column-wise variance}

\item{trans}{logic, transform Matrix Normal parameters into standard form, otherwise just display}
}
\value{
vector
}
\description{
transfrorm {alpha, Beta, Sigma} into {logit(alpha/2), Beta, SLR(Sigma)}, where 
SLR(Sigma) = {D = (var1 + var2)/2, D2 = (var1 - var2), logit((rho+1)/2)}
}
