% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite-prof.R
\name{composite.prof}
\alias{composite.prof}
\title{Profile Composite Likelihood}
\usage{
composite.prof(suff, Y, X, acf, ds, noSigma = FALSE)
}
\arguments{
\item{Y}{Original time series following Matrix Normal MN(X*Beta, V, Sigma), downsampled case 
would be Y_ds ~ MN(X_ds*Beta, V_ds, Sigma)}

\item{X}{Linear drift of time series. If X is of length 1, X_ds = rep(X, n)}

\item{acf}{ACF of columnwise-variance matrix \code{V_ds}, either vector or Toeplitz-object}

\item{ds}{Downsampling rate, starting at 2}

\item{noSigma}{logic, assume Sigma is identity matrix if true.}
}
\value{
Composite likelihood likelihood using profile sufficient statistics
}
