% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite-suff.R
\name{composite.suff}
\alias{composite.suff}
\title{Sufficient Statistics for Profile Composite Likelihood Inference}
\usage{
composite.suff(Y, X, acf, ds)
}
\arguments{
\item{Y}{Original time series following Matrix Normal MN(X*Beta, V, Sigma), downsampled case
would be Y_ds ~ MN(X_ds*Beta, V_ds, Sigma)}

\item{X}{Linear drift of time series. If X is of length 1, X_ds = rep(X, n)}

\item{acf}{ACF of columnwise-variance matrix \code{V_ds}, either vector or Toeplitz-object}

\item{ds}{Downsampling rate, starting at 2}
}
\value{
a list with the following elements:
\itemize{
   \item \code{Betahat = sum(t(Xi)V^{-1}Xi)^{-1}sum(t(Xi)V^{-1}Yi})
   \item \code{S = sum(t(Yi-Xi*Betahat)V^{-1}(Yi-Xi*Betahat))/n.ds/ds}
   \item \code{ldV = log(|V|)}
   \item \code{n.ds = floor(nrow(Y)/ds) - 1}
}
}
\note{
length of acf should equals floor(Yt/ds) - 1
}
