% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csi_fit.R
\name{csi_fit}
\alias{csi_fit}
\title{Fitting location-scale model for Gaussian CSI process.}
\usage{
csi_fit(model, dX, dT, Tz, var_calc)
}
\arguments{
\item{model}{An S3 object of type `csi_class` (see \code{\link{fbm_model}}, \code{\link{floc_model}}, \code{\link{farma_model}})}

\item{dX}{One or two-column matrix of increments. Increments of Gaussian CSI processes.}

\item{dT}{Interobservation time \eqn{\Delta t =} 1/fps.}

\item{Tz}{Optional Toeplitz matrix for intermediate calculations.}

\item{var_calc}{If \code{TRUE}, also estimate variance matrix.}
}
\value{
Vector of coefficients and possibly variance matrix on the transformed scale (see Details).
}
\description{
Fitting location-scale model for Gaussian CSI process.
}
\details{
The location-scale model for CSI is of the form
\deqn{
\Delta X_n = \mu \Delta t + \Sigma^{1/2} \Delta Z_n,
}
where \eqn{\Delta Z_n = Z_n - Z_{n-1}} is the increment process of CSI process.
}
\examples{
model <- fbm_model() # generating fBM model object
fits <- csi_fit(model, dX, dT, Tz, var_calc = TRUE)

# generating time series
dT <- 1/60
N <- 1800
acf1 <- fbm_acf(0.8, dT, N)
dX <- rSnorm(n = 2, acf = acf1)
Tz <- Toeplitz(n = N)
}
