% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csi_resid.R
\name{csi_resid}
\alias{csi_resid}
\title{Residuals of a Gaussian CSI process.}
\usage{
csi_resid(dX, dt, mu, acf, Sigma)
}
\arguments{
\item{dX}{One or two-column matrix of increments. Increments of Gaussian CSI processes.}

\item{dt}{Interobservation time \eqn{\Delta t} = 1/fps (positive scalar).}

\item{mu}{Vector of linear drift in each direction.}

\item{acf}{Autocorrelation, a vector of length \code{nrow(dX)}.}

\item{Sigma}{Covariance matrix of size \verb{ncol(dX) x ncol(dX)}.}
}
\value{
A residual matrix of the same size as \code{dX}.
}
\description{
Residuals of a Gaussian CSI process.
}
\details{
The residuals are calculated as
\deqn{
Z = T^{-1/2} (\Delta X - \mu \Delta T) \Sigma^{-1/2}
}{
Z = T^{-1/2} (\Delta X - \mu \Delta T) \Sigma^{-1/2}
}
where \eqn{T} is the Toeplitz matrix whose first column is \code{acf}.

The "square roots" correspond to the Cholesky decomposition for the Toeplitz row-dependence matrix, and the eigen decomposition for the column-dependence matrix.
}
\examples{
# generating 2-d stationary time series
alpha <- 0.8
dt <- 1/60
N <- 1800
acf1 <- fbm_acf(alpha, dt, N)
dX <- rSnorm(2, acf = acf1)
# Toeplitz object for internal computation
Tz <- Toeplitz(n = N)
mu <- c(1,2)
Sigma <- diag(2)
# compute the residuals
csi_resid(dX, dt, mu, acf1, Sigma)
}
