% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csi_resid.R
\name{csi_resid}
\alias{csi_resid}
\title{Calculate the residuals of a CSI process.}
\usage{
csi_resid(dX, drift, acf, Sigma)
}
\arguments{
\item{dX}{Matrix of trajectory increments.}

\item{drift}{Matrix of the same size as \code{dX} of increment drift in each coordinate per timepoint (see 'Details').}

\item{acf}{Autocorrelation vector of length \code{nrow(dX)} (see 'Details').}

\item{Sigma}{Covariance matrix of size \verb{ncol(dX) x ncol(dX)}.}
}
\value{
A residual matrix of the same size as \code{dX}.
}
\description{
Calculate the residuals of a CSI process.
}
\details{
The residuals are calculated as\preformatted{Z = toeplitz(acf)^\{-1/2\} \%*\% (dX - drift) \%*\% Sigma^\{-1/2\},
}

where the "square roots" correspond to the Cholesky decomposition for the row-dependence matrix \code{toeplitz(acf)}, and the eigen decomposition for the column-dependence matrix \code{Sigma}.
}
\examples{
# generating 2-d stationary time series
alpha <- 0.8
dt <- 1/60
N <- 1800
acf1 <- fbm_acf(alpha, dt, N)
dX <- rSnorm(2, acf = acf1)
# Toeplitz object for internal computation
Tz <- Toeplitz(n = N)
mu <- c(1,2)
Sigma <- diag(2)
# compute the residuals
csi_resid(dX, drift = rep(dt, N) \%o\% mu, acf1, Sigma)
}
