% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample}
\alias{downsample}
\title{Downsample a multi-dimensional time series.}
\usage{
downsample(Xt, r, position = 1)
}
\arguments{
\item{Xt}{One or two-column matrix of observations.}

\item{r}{Integer of downsampling rate.}

\item{position}{Integer of position index for downsampled series, smaller than or equal to \code{r}.}
}
\value{
An matrix of downsampled observation.
}
\description{
Downsample a multi-dimensional time series.
}
\details{
Sample observations from original time series \eqn{X = [X_0, X_2, X_3, ..., X_N]} at constant rate \code{r}. For given position index \code{i} finally returns \eqn{X_{ds} = [X_i, X_{i+r}, X_{i+2r}, ...]}.
}
\examples{
# simulate a 2D fbm trajectory.
alpha <- 0.8
dt <- 1/60
N <- 1800
acf <- fbm_acf(alpha, dt, N)
dX <- SuperGauss::rnormtz(n = 2, acf = acf)
Xt <- apply(dX, 2, cumsum)

# downsample Xt by rate 2.
Xt_ds2 <- downsample(Xt, r = 2)
}
