% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farma_acf.R
\name{farma_acf}
\alias{farma_acf}
\title{ACF of the Autoregressive Moving Average Model with fBM Noise}
\usage{
farma_acf(alpha, theta, rho, dT, N, m = 30)
}
\arguments{
\item{alpha}{TBD}

\item{theta}{parameter of AR model}

\item{rho}{parameter of MA model}

\item{dT}{TBD}

\item{N}{TBD}
}
\value{
Vector of length N
}
\description{
ACF of the Autoregressive Moving Average Model with fBM Noise
}
\details{
\code{Yt = theta Yt-1 + rho0 Xt + ...}, where \code{Xt} is fBM process with parameter \code{alpha}.
}
