% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fARMA_acf.R
\name{farma_acf}
\alias{farma_acf}
\title{ACF of the Autoregressive Moving Average Model with fBM Noise}
\usage{
farma_acf(alpha, gamma, rho, dT, N, p = 50)
}
\arguments{
\item{alpha}{TBD}

\item{gamma}{parameter of AR model}

\item{rho}{parameter of MA model}

\item{dT}{TBD}

\item{N}{TBD}
}
\value{
Vector of length N
}
\description{
ACF of the Autoregressive Moving Average Model with fBM Noise
}
\details{
\code{Yt = gamma Yt-1 + rho Xt + (1-gamma-rho) Xt-1}, where \code{Xt} is fBM process with parameter \code{alpha}.
}
