% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farma_acf.R
\name{farma_acf}
\alias{farma_acf}
\title{Calculate the autocorrelation function for the farma(p,q) model.}
\usage{
farma_acf(alpha, phi, rho, dT, N, m = 30)
}
\arguments{
\item{alpha}{Power law exponent of fBM process. A scalar between 0 and 2.}

\item{phi}{A vector of AR coefficients.}

\item{rho}{A vector of MA coefficients.}

\item{dT}{Interobservation time \eqn{\Delta t =} 1/fps (positive scalar).}

\item{N}{Length of observations (positive integer).}

\item{m}{Number of MA coefficients used for approximating AR coefficients (see \strong{Details}).}
}
\value{
A vector of \code{N} autocorrelation values.
}
\description{
Compute the autocorrelation of increments of farma(p,q) model (see \strong{Details}).
}
\details{
The farma(p,q) model is of following form:
\deqn{
Y_t = \sum_{i=1}^p \phi_i Y_{t-i} + \sum_{j=0}^q \rho_j X_{t-j}, \rho_0 = 1-\sum_{i=1}^p \phi_i-\sum_{j=1}^q \rho_j,
}{

}
where residuals \eqn{X_t} follows the fBM model with parameter \code{alpha}.
Auto-regressive terms are approximated with \code{m} terms of Moving-average terms.
}
\examples{
acf1 <- farma_acf(alpha = 0.8, phi = 0.1, rho = c(0.1, 0.05), dT = 1/60, N = 1800)
}
