% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farma_fit.R
\name{farma_fit}
\alias{farma_fit}
\title{Fit the fractional ARMA(1,1) model.}
\usage{
farma_fit(dX, dT, Tz, alpha, var_calc = TRUE, ...)
}
\arguments{
\item{dX}{One or two-column matrix of increments.}

\item{dT}{Interobservation time.}

\item{Tz}{Optional Toeplitz matrix for intermediate calculations.}

\item{alpha}{Subdiffusion parameter, optional.}

\item{var_calc}{If \code{TRUE}, also estimate variance matrix.}

\item{...}{Additional arguments to \code{\link[stats]{optim}}.}
}
\value{
Vector of coefficients and possibly variance matrix on the transformed scale (see Details).
}
\description{
Fit the fractional ARMA(1,1) model.
}
\details{
The fractional ARMA(1,1) model has the form
\deqn{
\Delta X_n = \gamma \Delta X_{n-1} + \rho \Delta Z_n + (1-\theta-\rho)  \Delta Z_{n-1},
}
where \eqn{\Delta Z_n} are increments of a 1D or 2D fBM process. The MLE and variance estimate are calculated on the transformed scale defined by \code{trans(rho) = logit(1-rho/2)}, \code{trans(mu) = mu}, \code{\link{trans_alpha}}, and \code{\link{trans_Sigma}}.
}
