% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farma_fit.R
\name{farma_fit}
\alias{farma_fit}
\title{Fit the farma model.}
\usage{
farma_fit(dX, dT, order, Tz, var_calc = TRUE)
}
\arguments{
\item{dX}{One or two-column matrix of increments. Increments of Gaussian CSI processes.}

\item{dT}{Interobservation time \eqn{\Delta t} = 1/fps (positive scalar).}

\item{order}{A specification of the farma model: the two integer components (p, q) are the AR order and the MA order.}

\item{Tz}{Optional object of class \code{Toeplitz} for intermediate calculations (See \code{\link[SuperGauss:Toeplitz]{SuperGauss::Toeplitz()}}).}

\item{var_calc}{Logical; if \code{TRUE}, also estimate the variance matrix.}
}
\value{
A vector of estimated parameters on transformed scale (See \code{\link[=farma_model]{farma_model()}}). If \code{var_calc}, a list with components:
\describe{
\item{coef}{A vector of estimated parameters on transformed scale.}
\item{vcov}{A matrix of estimated covariance of parameters on transformed scale.}
}
}
\description{
Fit a farma(p,q) model to a multi-dimensional CSI process (See \strong{Details}).
}
\details{
The farma(p,q) model is of following form:
\deqn{
Y_n = \sum_{i=1}^p \phi_i Y_{n-i} + \sum_{j=0}^q \rho_j X_{n-j}
}{
Y[n] = \phi_1 Y_(n-1) + ... + \phi_p Y_(n-p) + \rho_0 X_(n) + ... + \rho_q X_(n-q)
}
where \eqn{X_n} is an \verb{q = 1,2} dimensional fBM model (See \code{\link[=fbm_fit]{fbm_fit()}}).

Optimization is done by \code{stats::optimize}. It works better when parameters are re-parametrized into unrestricted form (See \code{\link[=farma_model]{farma_model()}}).
}
\examples{
# generating 2-d stationary time series
alpha <- 0.8
dT <- 1/60
N <- 1800
acf1 <- fbm_acf(alpha, dT, N)
dX <- rSnorm(2, acf = acf1)
# Toeplitz object for internal computation
Tz <- Toeplitz(n = N)
# Fit the farma(0,1) model
farma_fit(dX, dT, order = c(0,1), Tz, var_calc = TRUE)
}
