% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farma_fit.R
\name{farma_fit}
\alias{farma_fit}
\title{Fit the fractional Moving-average model.}
\usage{
farma_fit(dX, dT, order, Tz, var_calc = TRUE)
}
\arguments{
\item{dX}{One or two-column matrix of increments.}

\item{dT}{Interobservation time.}

\item{Tz}{Optional Toeplitz matrix for intermediate calculations.}

\item{var_calc}{If \code{TRUE}, also estimate variance matrix.}

\item{nlag}{Number of lags (see Details).}

\item{alpha}{Subdiffusion parameter, optional.}

\item{...}{Additional arguments to \code{\link[stats]{optim}}.}
}
\value{
Vector of coefficients and possibly variance matrix on the transformed scale (see Details).
}
\description{
Fit the fractional Moving-average model.
}
\details{
The fractional Moving-average model has the form
\deqn{
\Delta X_n = (1-\sum_{i=1}^p \rho_i) \Delta Z_n + \sum_{i=1}^p \rho_{i} \Delta Z_{n-i},
}
where \eqn{\Delta Z_n} are increments of a 1D or 2D fBM process. The MLE and variance estimate are calculated on the transformed scale defined by \code{trans(rho) = logit(1-rho/2)}, \code{trans(mu) = mu}, \code{\link{trans_alpha}}, and \code{\link{trans_Sigma}}.
}
