% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farma_fit.R
\name{farma_fit}
\alias{farma_fit}
\title{Fit the ARMA(p,q) filter with fBM residuals.}
\usage{
farma_fit(dX, dT, order, Tz, var_calc = TRUE)
}
\arguments{
\item{dX}{One or two-column matrix of increments. Increments of Gaussian CSI processes.}

\item{dT}{Interobservation time \eqn{\Delta t =} 1/fps (positive scalar).}

\item{order}{A specification of the ARMA filter: the two integer components (p, q) are the AR order and the MA order.}

\item{Tz}{Optional Toeplitz matrix for intermediate calculations.}

\item{var_calc}{If \code{TRUE}, also estimate variance matrix.}
}
\value{
Vector of coefficients and possibly variance matrix on the transformed scale (see Details).
}
\description{
Fit the ARMA(p,q) filter with fBM residuals.
}
\details{
The ARMA(p,q) filter is of the form
\deqn{
Y_t = \sum_{i=1}^p \phi_i Y_{t-i} + \sum_{j=0}^q \rho_j X_{t-j}, \rho_0 = 1-\sum_{i=1}^p \phi_i-\sum_{j=1}^q \rho_j,
}
where residuals \eqn{X_t} follows the fBM model:
\deqn{
X_t = \mu t + \Sigma^{1/2} Z_t,
}
where \eqn{Z_t} consists of \code{q = 1,2} iid fBM processes.
}
\examples{
fits <- farma_fit(dX, dT, order = c(1,1), Tz, var_calc = TRUE) # fitting ARMA(1,1) filter

# generating time series
dT <- 1/60
N <- 1800
acf1 <- fbm_acf(0.8, dT, N)
dX <- rSnorm(n = 2, acf = acf1)
Tz <- Toeplitz(n = N)
}
