% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farma_model.R
\name{farma_model}
\alias{farma_model}
\title{Constructor for fARMA model object.}
\usage{
farma_model(p, q)
}
\arguments{
\item{p}{Integer of Auto-regressive order.}

\item{q}{Integer of Moving-average order.}
}
\value{
An S3 object of type `csi_class`, consisting of a list wuth elements:
\itemize{
\item theta_names: A string vector of parameter names.
\item acf: ACF function with three arguments: vector of ACF parameters \eqn{\theta}; interobservation time \eqn{\Delta t = 1/fps}; length \eqn{N}.
\item theta_trans: Transformation function that converts constrained ACF parameter \eqn{\theta} into unrestricted parameters \eqn{\gamma}.
\item theta_itrans: Inverse Transformation function that converts unrestricted parameters \eqn{\gamma} into constrained ACF parameter \eqn{\theta}.
\item penalty: Penalty function for unrestricted parameters \eqn{\gamma}.
}
}
\description{
Constructor for fARMA model object.
}
\details{
Constructor function that generates the default S3 class of fARMA model.
}
\examples{
model <- farma_model(1,1)

}
