% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbm_acf.R
\name{fbm_acf}
\alias{fbm_acf}
\title{Acf function for the increment of fBM process.}
\usage{
fbm_acf(alpha, dT, N)
}
\arguments{
\item{alpha}{fBM parameter, range from 0 to 2.}

\item{dT}{interobservation time.}

\item{N}{number of samples.}
}
\value{
vector of length N.
}
\description{
Acf function for the increment of fBM process.
}
