% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbm_acf.R
\name{fbm_acf}
\alias{fbm_acf}
\title{Calculate the ACF for the fBM model.}
\usage{
fbm_acf(alpha, dT, N)
}
\arguments{
\item{alpha}{Power law exponent. A scalar between 0 and 2.}

\item{dT}{Interobservation time \eqn{\Delta t =} 1/fps (positive scalar).}

\item{N}{Length of observations (positive integer).}
}
\value{
A vector of \code{N} autocorrelation values.
}
\description{
Compute the autocovariance of fBM increments of equally-spaced observations (see \strong{Details}).
}
\details{
The autocovariance of fBM increments at lag \eqn{h} is given by
\deqn{
\textrm{acf}(h) = \Delta t^\alpha/2 \times \left(|h+1|^\alpha + |h-1|^\alpha - 2|h|^\alpha\right).
}{
acf(h) = \Delta t^\alpha/2 * (|h+1|^\alpha + |h-1|^\alpha - 2|h|^\alpha). 
}

This function returns the autocovariance function of increments of fBM process.
}
\examples{
# ACF parameters
alpha <- 0.8
dT <- 1/60
N <- 1800
# ACF of fBM model
fbm_acf(alpha, dT, N)
}
