% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbm_fit.R
\name{fbm_fit}
\alias{fbm_fit}
\title{Fit the fBM model.}
\usage{
fbm_fit(dX, dT, Tz, var_calc = TRUE)
}
\arguments{
\item{dX}{One or two-column matrix of increments.}

\item{dT}{Interobservation time.}

\item{Tz}{Optional Toeplitz matrix for intermediate calculations.}

\item{var_calc}{If \code{TRUE}, also estimate variance matrix.}
}
\value{
Vector of coefficients and possibly variance matrix on the transformed scale (see Details).
}
\description{
Fit the fBM model.
}
\details{
The fBM model is of the form
\deqn{
X_t = \mu t + \Sigma^{1/2} Z_t,
}
where \eqn{Z_t} consists of \code{q = 1,2} iid fBM processes with \eqn{MSD_Z(t) = t^\alpha}.  The MLE and variance estimate are calculated on the transformed scale defined by \code{trans(mu) = mu}, \code{\link{trans_alpha}}, and \code{\link{trans_Sigma}}.
}
