% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbm_fit.R
\name{fbm_fit}
\alias{fbm_fit}
\title{Fit the fBM model.}
\usage{
fbm_fit(
  Xt,
  dt,
  drift = c("linear", "none", "quadratic"),
  vcov = TRUE,
  ad_only = TRUE
)
}
\arguments{
\item{Xt}{Matrix of trajectory positions, where each row is an observation and each column is a measurement coordinate.  The trajectory is assumed to be sampled at a constant frequency.}

\item{dt}{Interobservation time \eqn{\Delta t} = 1/fps (positive scalar).}

\item{drift}{Character string specifying one of the \emph{preset} drift types "linear", "none", or "quadratic".  Custom drift functions are not supported through this simplified interface.  See \link{csi_model} for details.}

\item{ad_only}{Whether to return estimates of (\verb{alpha, log(D)}) only, or the entire parameter vector \code{omega} in the computational basis.}

\item{var_calc}{Logical; if \code{TRUE}, also estimate the variance matrix.}
}
\value{
A vector of estimated parameters on transformed scale (See \code{\link[=fbm_model]{fbm_model()}}). If \code{vcov}, a list with components:
\describe{
\item{coef}{A vector of estimated parameters on transformed scale.}
\item{vcov}{A matrix of estimated covariance of parameters on transformed scale.}
}
}
\description{
Fit a fBM model to a multi-dimensional CSI process.
}
\details{
The fBM model is of the form:
\deqn{
X_t = \mu t + \Sigma^{1/2} Z_t
}{
X[t] = \mu t + \Sigma^{1/2} Z[t],
}
where \eqn{Z[t]} consists of \verb{q = 1,2} iid fBM processes with \eqn{MSD_Z(t) = t^\alpha}.

Optimization is done by \code{\link[stats:optimize]{stats::optimize()}}. It works better when parameters are re-parametrized into unrestricted form (See \code{\link[=fbm_model]{fbm_model()}}).
}
\examples{
# generating 2-d stationary time series
alpha <- 0.8
dt <- 1/60
N <- 1800
acf1 <- fbm_acf(alpha, dt, N)
dX <- rSnorm(2, acf = acf1)
# Toeplitz object for internal computation
Tz <- Toeplitz(n = N)
# Fit the fBM model
fbm_fit(dX, dt, Tz, vcov = TRUE)
}
