% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbm_fit.R
\name{fbm_fit}
\alias{fbm_fit}
\title{Fit the fBM model.}
\usage{
fbm_fit(dX, dT, Tz, var_calc = TRUE)
}
\arguments{
\item{dX}{One or two-column matrix of increments. Increments of Gaussian CSI processes.}

\item{dT}{Interobservation time \eqn{\Delta t =} 1/fps (positive scalar).}

\item{Tz}{Optional Toeplitz matrix for intermediate calculations.}

\item{var_calc}{If \code{TRUE}, also estimate variance matrix.}
}
\value{
Vector of coefficients and possibly variance matrix on the transformed scale (see Details).
}
\description{
Fit the fBM model.
}
\details{
The fBM model is of the form
\deqn{
X_t = \mu t + \Sigma^{1/2} Z_t,
}
where \eqn{Z_t} consists of \code{q = 1,2} iid fBM processes with \eqn{MSD_Z(t) = t^\alpha}.
}
\examples{
# generating time series
dT <- 1/60
N <- 1800
acf1 <- fbm_acf(0.8, dT, N)
dX <- rSnorm(n = 2, acf = acf1)
Tz <- Toeplitz(n = N)
}
