% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbm_model.R
\name{fbm_model}
\alias{fbm_model}
\title{Class definition for the fBM model.}
\description{
Class definition for the fBM model.
}
\details{
Fractional Brownian motion \code{X_t} is the only (zero-mean) continuous stationary increments (CSI) Gaussian process having mean square displacement (MSD) function given by the power law

\if{html}{\out{<div class="sourceCode">}}\preformatted{E[(X_t-X_0)^2] = t^alpha,
}\if{html}{\out{</div>}}

with subdiffusion exponent \verb{0 < alpha < 2}.  The resulting MSD as a function of \code{alpha} is what gets passed to the \link{csi_model} base class to construct the \code{fbm_model} derived class.
}
\examples{
# simulate data from the fbm model
alpha <- .8

dt <- 1/60
N <- 1800
ndim <- 2

Xt <- csi_sim(drift = matrix(0, N-1, ndim),
              acf = fbm_acf(alpha, dt, N-1),
              Sigma = diag(ndim),
              X0 = rep(0, ndim))


# create fbm model object
model <- fbm_model$new(Xt = Xt, dt = dt, drift = "linear")

# evaluate loglikelihood
model$loglik(phi = c(alpha = alpha),
             mu = rep(0, ndim),
             Sigma = diag(ndim))

}
\section{Super class}{
\code{\link[subdiff:csi_model]{subdiff::csi_model}} -> \code{fbm_model}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{phi_names}}{Kernel parameter names. In this case, the character string \code{alpha}.  See \link{csi_model}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-fbm_model-acf}{\code{fbm_model$acf()}}
\item \href{#method-fbm_model-trans}{\code{fbm_model$trans()}}
\item \href{#method-fbm_model-itrans}{\code{fbm_model$itrans()}}
\item \href{#method-fbm_model-get_subdiff}{\code{fbm_model$get_subdiff()}}
\item \href{#method-fbm_model-clone}{\code{fbm_model$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="drift"><a href='../../subdiff/html/csi_model.html#method-csi_model-drift'><code>subdiff::csi_model$drift()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="fisher"><a href='../../subdiff/html/csi_model.html#method-csi_model-fisher'><code>subdiff::csi_model$fisher()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="fit"><a href='../../subdiff/html/csi_model.html#method-csi_model-fit'><code>subdiff::csi_model$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="get_omega"><a href='../../subdiff/html/csi_model.html#method-csi_model-get_omega'><code>subdiff::csi_model$get_omega()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="get_vcov"><a href='../../subdiff/html/csi_model.html#method-csi_model-get_vcov'><code>subdiff::csi_model$get_vcov()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="initialize"><a href='../../subdiff/html/csi_model.html#method-csi_model-initialize'><code>subdiff::csi_model$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="itrans_full"><a href='../../subdiff/html/csi_model.html#method-csi_model-itrans_full'><code>subdiff::csi_model$itrans_full()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="loglik"><a href='../../subdiff/html/csi_model.html#method-csi_model-loglik'><code>subdiff::csi_model$loglik()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="msd"><a href='../../subdiff/html/csi_model.html#method-csi_model-msd'><code>subdiff::csi_model$msd()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="nlp"><a href='../../subdiff/html/csi_model.html#method-csi_model-nlp'><code>subdiff::csi_model$nlp()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="nu_hat"><a href='../../subdiff/html/csi_model.html#method-csi_model-nu_hat'><code>subdiff::csi_model$nu_hat()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="resid"><a href='../../subdiff/html/csi_model.html#method-csi_model-resid'><code>subdiff::csi_model$resid()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="sim"><a href='../../subdiff/html/csi_model.html#method-csi_model-sim'><code>subdiff::csi_model$sim()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="trans_full"><a href='../../subdiff/html/csi_model.html#method-csi_model-trans_full'><code>subdiff::csi_model$trans_full()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fbm_model-acf"></a>}}
\if{latex}{\out{\hypertarget{method-fbm_model-acf}{}}}
\subsection{Method \code{acf()}}{
Increment autocorrelation function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fbm_model$acf(phi, dt, N)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phi, dt, N}}{See \link{csi_model}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fbm_model-trans"></a>}}
\if{latex}{\out{\hypertarget{method-fbm_model-trans}{}}}
\subsection{Method \code{trans()}}{
Transform kernel parameters from regular to computational basis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fbm_model$trans(phi)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phi}}{See \link{csi_model}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fbm_model-itrans"></a>}}
\if{latex}{\out{\hypertarget{method-fbm_model-itrans}{}}}
\subsection{Method \code{itrans()}}{
Transform kernel parameters from computational to regular basis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fbm_model$itrans(psi)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{psi}}{See \link{csi_model}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fbm_model-get_subdiff"></a>}}
\if{latex}{\out{\hypertarget{method-fbm_model-get_subdiff}{}}}
\subsection{Method \code{get_subdiff()}}{
Transform parameters from computational basis to subdiffusion parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fbm_model$get_subdiff(omega)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{omega}}{See \link{csi_model}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector with named elements \code{alpha} and \code{logD}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fbm_model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-fbm_model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fbm_model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
