% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbm_model.R
\name{fbm_model}
\alias{fbm_model}
\title{Constructor of fBM model object.}
\usage{
fbm_model()
}
\value{
An list of class \code{csi_class} with components:
\describe{
\item{\code{theta_names}}{A vector of parameter names.}
\item{\code{acf}}{An ACF function with three arguments: vector of parameters; interobservation time \eqn{\Delta t}; length \code{N}.}
\item{\code{theta_trans}}{Transformation function that converts original ACF parameter into unrestricted parameters.}
\item{\code{theta_itrans}}{Inverse Transformation function that converts unrestricted parameters into original ACF parameter.}
\item{\code{penalty}}{Penalty function for unrestricted parameters.}
}
}
\description{
Constructor of fBM model object.
}
\details{
Constructor function that generates a list of fBM model for purpose of fitting.
}
\examples{
# generate a fBM model object
model <- fbm_model()
# we can modify the elements of generated object
model$theta_names <- "gamma"
}
