% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floc_fit.R
\name{fdl_fit}
\alias{fdl_fit}
\alias{floc_fit}
\title{Fit the fBM model with dynamic and static errors.}
\usage{
floc_fit(dX, dT, Tz, var_calc = TRUE)
}
\arguments{
\item{dX}{One or two-column matrix of increments.}

\item{dT}{Interobservation time.}

\item{Tz}{Optional Toeplitz matrix for intermediate calculations.}

\item{var_calc}{If \code{TRUE}, also estimate variance matrix.}

\item{tau}{Scalar between 0 and 1, indicating the percentage of time for which the camera shutter is open in the dynamic error model.  Estimated if missing. See Details.}

\item{sigma2}{Magnitude of static error.  Estimated if missing. See Details.}

\item{penalty}{logic, employ a small penalty on \code{(tau, sigma2)} when \code{TRUE}}
}
\value{
Vector of coefficients and possibly variance matrix on the transformed scale (see Details).
}
\description{
Fit the fBM model with dynamic and static errors.
}
\details{
The fBM + dynamic and localization error (fdl) model has the form
\deqn{
X_n = 1/\tau \int_0^\tau Z_(n+s) ds + \sigma e_{n},
}
where \eqn{Z_n} is a 1D or 2D fBM process. The MLE and variance estimate are calculated on the transformed scale defined by \code{trans(tau) = logit(tau)}, \code{trans(sigma) = log(sigma)}, \code{trans(mu) = mu}, \code{\link{trans_alpha}}, and \code{\link{trans_Sigma}}.
When put \code{tau} = 0, this model becomes fractional localization error model.
When put \code{sigma2} = 0, this model becomes fractional dynamic error model.
When put \code{(tau, sigma2)} = 0, this model becomes fractional Brownian model.
}
