% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdl_fit.R
\name{fdl_fit}
\alias{fdl_fit}
\title{Fit the fractional dynamic localization error model.}
\usage{
fdl_fit(dX, dT, Tz, type = "dynamic localization", var_calc = TRUE, ...)
}
\arguments{
\item{dX}{one or two-column matrix of trajectory increments.}

\item{dT}{Interobservation time.}

\item{Tz}{Optional Toeplitz matrix for intermediate calculations.}

\item{type}{One of "dynamic localization", "dynamic" or "localization".}

\item{var_calc}{If \code{TRUE}, also estimate variance matrix.}

\item{...}{Additional \code{control} arguments to \code{stats::optim}.}
}
\value{
Vector of coefficients and possibly variance matrix on the transformed scale (see Details).
}
\description{
Fit the fractional dynamic localization error model.
}
\details{
The fractional AR(1) model has the form
\deqn{
X_n = 1/\tau \int_0^\tau Z_(n+s) ds + \sigma e_{n},
}
where \eqn{Z_n} is a 1D or 2D fBM process. The MLE and variance estimate are calculated on the transformed 
scale defined by \code{trans(rho) = logit(1-rho/2)}, \code{trans(mu) = mu}, \code{\link{trans_alpha}}, and 
\code{\link{trans_Sigma}}.
}
