% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fds_fit.R
\name{fds_fit}
\alias{fds_fit}
\title{fBM inference using composite downsampling}
\usage{
fds_fit(dX, dT, ds, var_calc = TRUE, Tz)
}
\arguments{
\item{dX}{one or two-column matrix of trajectory increments.}

\item{dT}{Interobservation time.}

\item{var_calc}{If \code{TRUE}, also estimate variance matrix.}

\item{Tz}{Optional Toeplitz matrix for intermediate calculations.}

\item{type}{type of downsampling method, "naive" means using basic downsampling method, "comp"
means using composite downsampling method}

\item{...}{Additional \code{control} arguments to \code{stats::optim}.}
}
\value{
Vector of coefficients and possibly variance matrix on the transformed scale (see Details).
}
\description{
fBM inference using composite downsampling
}
\details{
We assume that time series dX follows MN(Beta * dT, V_fBM(alpha, dT), Sigma),
thus downsampled dY_ds follows MN(Beta * ds * dT, V_fBM(alpha, dT*ds), Sigma).
}
