% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floc_acf.R
\name{floc_acf}
\alias{floc_acf}
\title{ACF of fBM + Savin & Doyle's Localization Error.}
\usage{
floc_acf(alpha, tau, sigma2, dT, N)
}
\arguments{
\item{alpha}{Parameter of fBM process.}

\item{tau}{Real number ranging between 0 and 1, indicating the percentage of time for which the camera shutter is open in the dynamic error model. See Details.}

\item{sigma2}{Magnitude of static error. See Details.}

\item{dT}{Interobservation time \eqn{\Delta t =} 1/fps (positive scalar).}

\item{N}{Length of observations (positive integer).}
}
\value{
A vector of \code{N} autocorrelation values.
}
\description{
ACF of fBM + Savin & Doyle's Localization Error.
}
\details{
this function returns the ACF of \eqn{\Delta Y_n}, where \eqn{Y_n} follows fLOC model:
\deqn{
Y_n = 1/\tau \int_0^\tau X_{n+s} ds + \sigma e_{n},
},
where \eqn{X_n} is an fBM process and \eqn{e_n} is a white noise with variance \code{sigma2}.
}
\examples{
floc_acf(alpha = 0.8, tau = 1/60, sigma2 = 0.01, dT = 1/60, N = 1800)

}
