% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floc_model.R
\name{floc_model}
\alias{floc_model}
\title{Constructor of fLOC model object.}
\usage{
floc_model()
}
\value{
An list of class `csi_class` with components:
\describe{
\item{theta_names}{A vector of parameter names.}
\item{acf}{An ACF function with three arguments: vector of parameters; interobservation time \eqn{\Delta t}; length \code{N}.}
\item{theta_trans}{Transformation function that converts original ACF parameter into unrestricted parameters.}
\item{theta_itrans}{Inverse Transformation function that converts unrestricted parameters into original ACF parameter.}
\item{penalty}{Penalty function for unrestricted parameters.}
}
}
\description{
Constructor of fLOC model object.
}
\details{
Constructor function that generates a list of floc model for purpose of fitting.
}
\examples{
# generate a fLOC model object
model <- floc_model()
# we can modify the elements of generated object
model$theta_names[1] <- "gamma"
}
