% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farma_acf.R
\name{fma_acf}
\alias{fma_acf}
\title{ACF of unparametrized moving-average model with fBM noises.}
\usage{
fma_acf(alpha, rho, dT, N)
}
\description{
ACF of unparametrized moving-average model with fBM noises.
}
