% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fma_fit.R
\name{fma_fit}
\alias{fma_fit}
\title{Fit the fractional MA(1) model.}
\usage{
fma_fit(dX, dT, nlag, Tz, var_calc = TRUE, ...)
}
\arguments{
\item{dX}{one or two-column matrix of trajectory increments.}

\item{dT}{Interobservation time.}

\item{nlag}{Maximum number of lags for the fMA model.}

\item{Tz}{Optional Toeplitz matrix for intermediate calculations.}

\item{var_calc}{If \code{TRUE}, also estimate variance matrix.}

\item{...}{Additional arguments to \code{stats::optim}.}
}
\value{
Vector of coefficients and possibly variance matrix on the transformed scale (see Details).
}
\description{
Fit the fractional MA(1) model.
}
\details{
The fractional MA(1) model has the form
\deqn{
\Delta X_n = (1-\rho) \Delta Z_n + \rho \Delta Z_{n-1},
}
where \eqn{\Delta Z_n} are increments of a 1D or 2D fBM process. The MLE and variance estimate are calculated on the transformed scale defined by \code{trans(rho) = logit(1-rho/2)}, \code{trans(mu) = mu}, \code{\link{trans_alpha}}, and \code{\link{trans_Sigma}}.
}
