% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsd_model.R
\name{fsd_model}
\alias{fsd_model}
\title{Class definition for the fSD model.}
\description{
Class definition for the fSD model.
}
\details{
Let \code{X_t} denote the true position of an fBM process at time \code{t}.  The Savin-Doyle localization error model describes the measured position \code{Y_n} at time \verb{t = n dt} as

\if{html}{\out{<div class="sourceCode">}}\preformatted{Y_n = sigma * eps_n + 1/(tau*dt) * int_0^(tau*dt) X_\{n dt - s\} ds,
}\if{html}{\out{</div>}}

where \verb{eps_n ~iid N(0,1)} is a Gaussian white noise process.  The resulting MSD as a function of \code{tau} and \code{sigma2 = sigma^2} is what gets passed to the \link{csi_model} base class to construct the \code{fsd_model} derived class.
}
\examples{
# simulate data from the fsd model
alpha <- .8
tau <- .1
sigma2 <- .01

dt <- 1/60
N <- 1800
ndim <- 2

Xt <- csi_sim(drift = matrix(0, N-1, ndim),
              acf = fsd_acf(alpha, tau, sigma2, dt, N-1),
              Sigma = diag(ndim),
              X0 = rep(0, ndim))


# create fsd model object
model <- fsd_model$new(Xt = Xt, dt = dt, drift = "linear")

# evaluate loglikelihood
model$loglik(phi = c(alpha = alpha, tau = tau, sigma2 = sigma2),
             mu = rep(0, ndim),
             Sigma = diag(ndim))

}
\section{Super class}{
\code{\link[subdiff:csi_model]{subdiff::csi_model}} -> \code{fsd_model}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{phi_names}}{Kernel parameter names.  A subset of \verb{(alpha, tau, sigma2)} (see \link{csi_model} and \code{fsd_model$initialize()}).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-fsd_model-acf}{\code{fsd_model$acf()}}
\item \href{#method-fsd_model-trans}{\code{fsd_model$trans()}}
\item \href{#method-fsd_model-itrans}{\code{fsd_model$itrans()}}
\item \href{#method-fsd_model-get_subdiff}{\code{fsd_model$get_subdiff()}}
\item \href{#method-fsd_model-clone}{\code{fsd_model$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="drift"><a href='../../subdiff/html/csi_model.html#method-csi_model-drift'><code>subdiff::csi_model$drift()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="fisher"><a href='../../subdiff/html/csi_model.html#method-csi_model-fisher'><code>subdiff::csi_model$fisher()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="fit"><a href='../../subdiff/html/csi_model.html#method-csi_model-fit'><code>subdiff::csi_model$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="get_omega"><a href='../../subdiff/html/csi_model.html#method-csi_model-get_omega'><code>subdiff::csi_model$get_omega()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="get_vcov"><a href='../../subdiff/html/csi_model.html#method-csi_model-get_vcov'><code>subdiff::csi_model$get_vcov()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="initialize"><a href='../../subdiff/html/csi_model.html#method-csi_model-initialize'><code>subdiff::csi_model$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="itrans_full"><a href='../../subdiff/html/csi_model.html#method-csi_model-itrans_full'><code>subdiff::csi_model$itrans_full()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="loglik"><a href='../../subdiff/html/csi_model.html#method-csi_model-loglik'><code>subdiff::csi_model$loglik()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="msd"><a href='../../subdiff/html/csi_model.html#method-csi_model-msd'><code>subdiff::csi_model$msd()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="nlp"><a href='../../subdiff/html/csi_model.html#method-csi_model-nlp'><code>subdiff::csi_model$nlp()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="nu_hat"><a href='../../subdiff/html/csi_model.html#method-csi_model-nu_hat'><code>subdiff::csi_model$nu_hat()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="resid"><a href='../../subdiff/html/csi_model.html#method-csi_model-resid'><code>subdiff::csi_model$resid()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="sim"><a href='../../subdiff/html/csi_model.html#method-csi_model-sim'><code>subdiff::csi_model$sim()</code></a></span></li>
<li><span class="pkg-link" data-pkg="subdiff" data-topic="csi_model" data-id="trans_full"><a href='../../subdiff/html/csi_model.html#method-csi_model-trans_full'><code>subdiff::csi_model$trans_full()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fsd_model-acf"></a>}}
\if{latex}{\out{\hypertarget{method-fsd_model-acf}{}}}
\subsection{Method \code{acf()}}{
Increment autocorrelation function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fsd_model$acf(phi, dt, N)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phi, dt, N}}{See \link{csi_model}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fsd_model-trans"></a>}}
\if{latex}{\out{\hypertarget{method-fsd_model-trans}{}}}
\subsection{Method \code{trans()}}{
Transform kernel parameters from regular to computational basis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fsd_model$trans(phi)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{phi}}{See \link{csi_model}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The transformation function is

\if{html}{\out{<div class="sourceCode">}}\preformatted{psi = (psi1 = logit(alpha/2), psi2 = logit(tau), psi3 = log(sigma2)).
}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fsd_model-itrans"></a>}}
\if{latex}{\out{\hypertarget{method-fsd_model-itrans}{}}}
\subsection{Method \code{itrans()}}{
Transform kernel parameters from computational to regular basis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fsd_model$itrans(psi)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{psi}}{See \link{csi_model}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fsd_model-get_subdiff"></a>}}
\if{latex}{\out{\hypertarget{method-fsd_model-get_subdiff}{}}}
\subsection{Method \code{get_subdiff()}}{
Transform parameters from computational basis to subdiffusion parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fsd_model$get_subdiff(omega)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{omega}}{See \link{csi_model}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector with named elements \code{alpha} and \code{logD}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fsd_model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-fsd_model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{fsd_model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
