% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbe.R
\docType{data}
\name{hbe}
\alias{hbe}
\title{Simulated HBE dataset.}
\format{
A dataset containing 76 simulated 2D particle trajectories sampled at 60 Hz.  The four columns of the data are:
\describe{
\item{\code{id}}{Particle ID (1-76).}
\item{\code{t}}{Time at which position was recorded (in seconds).}
\item{\code{x}}{Particle position on \code{x}-axis at time \code{t} (in microns).}
\item{\code{y}}{Particle position on \code{y}-axis at time \code{t} (in microns).}
}
}
\usage{
hbe
}
\description{
Simulated HBE dataset.
}
\keyword{datasets}
