% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_fit.R
\name{ls_fit}
\alias{ls_fit}
\title{Fit least-squares estimate.}
\usage{
ls_fit(dX, dt, lags, type = c("standard", "improved"), vcov = TRUE)
}
\arguments{
\item{dX}{One or two-column matrix of increments. Increments of Gaussian CSI processes.}

\item{dt}{Interobservation time \eqn{\Delta t} = 1/fps (positive scalar).}

\item{lags}{Integer vector of lags to use in the fit, such that the timepoints used in the fit are \code{tau = dt * lags}.}

\item{type}{Either "standard" for the usual LS estimator, or "improved" for the version of Zhang et al (2018).}

\item{vcov}{If \code{TRUE}, calculates the variance matrix of \verb{(alpha, logD)}.}
}
\value{
If \code{vcov = FALSE}, vector of length 2 with estimates of \verb{(alpha, logD)}.  Otherwise, a list with elements \code{coef} and \code{vcov}, where the former is the estimate and the latter is the corresponding variance estimator.
}
\description{
Fit least-squares estimate.
}
\note{
Uses the \strong{subdiff} convention for \code{D}.
}
\references{
Zhang, K., Crizer, K.P.R., Schoenfisch, M.H., Hill, B.D., Didier, G. (2018) "Fluid heterogeneity detection based on the asymptotic distribution of the time-averaged mean squared displacement in single particle tracking experiments". \emph{Journal of Physics A: Mathematical and Theoretical}, 51, pp 445601(44).
}
