% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_fit.R
\name{ls_fit}
\alias{ls_fit}
\alias{ls_msd_fit}
\title{Calculate the least-squares estimate of \verb{(alpha, D)}.}
\usage{
ls_fit(Xt, dt, lags, type = c("standard", "improved"), vcov = TRUE)

ls_msd_fit(
  msd,
  dt,
  lags,
  N,
  ndim,
  type = c("standard", "improved"),
  vcov = TRUE
)
}
\arguments{
\item{Xt}{Matrix of trajectory positions, where each row is an observation and each column is a measurement coordinate.  The trajectory is assumed to be sampled at a constant frequency.}

\item{dt}{Interobservation time \eqn{\Delta t} = 1/fps (positive scalar).}

\item{lags}{Integer vector of lags to use in the fit, such that the timepoints used in the fit are \code{tau = dt * lags}.}

\item{type}{Either "standard" for the usual LS estimator, or "improved" for the version of Zhang et al (2018).}

\item{vcov}{If \code{TRUE}, returns an estimate of the variance matrix of \verb{(alpha, logD)}.}

\item{msd}{Vector of empirical MSD estimates computed by \code{\link[=msd_fit]{msd_fit()}} at the values of \code{lags}.}

\item{N}{Length of the particle trajectory.}

\item{ndim}{Number of dimensions of the particle trajectory.}
}
\value{
If \code{vcov = FALSE}, vector of length 2 with estimates of \verb{(alpha, logD)}.  Otherwise, a list with elements \code{coef} and \code{vcov}, where the former is the estimate and the latter is the corresponding variance estimator.
}
\description{
Calculate the least-squares estimate of \verb{(alpha, D)}.
}
\details{
\code{\link[=ls_fit]{ls_fit()}} first computes the MSD using \code{msd_fit(Xt, dt, demean = TRUE)} then passes this on to \code{\link[=ls_msd_fit]{ls_msd_fit()}}.  For finer control over the MSD or if it has been precomputed, one may interact with \code{\link[=ls_msd_fit]{ls_msd_fit()}} directly.
}
\note{
Uses the \strong{subdiff} convention for \code{D}.
}
\references{
Zhang, K., Crizer, K.P.R., Schoenfisch, M.H., Hill, B.D., Didier, G. (2018) "Fluid heterogeneity detection based on the asymptotic distribution of the time-averaged mean squared displacement in single particle tracking experiments". \emph{Journal of Physics A: Mathematical and Theoretical}, 51, pp 445601(44).
}
