% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_acf.R
\name{ma_acf}
\alias{ma_acf}
\title{ACF of the Moving Average Model}
\usage{
ma_acf(alpha, rho, dT, N)
}
\arguments{
\item{rho}{parameter of MA model}
}
\value{
vector of length N
}
\description{
ACF of the Moving Average Model
}
\details{
Yt = (1-rho) Xt + rho Xt-1, where Xt is fBM process with parameter alpha
}
