% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msd.R
\name{msd_fit}
\alias{msd_fit}
\title{Sample estimate of mean squared displacement.}
\usage{
msd_fit(Xt, nlag, dT = NULL)
}
\arguments{
\item{Xt}{Vector or matrix of trajectories (each one is a column).}

\item{nlag}{Number of msd lags to calculate.}

\item{dT}{Optional interobservation time.  If supplied, estimates subdiffusion parameters \code{alpha, D} by regressing MSD on log-log scale.}
}
\value{
Sample MSD vector of length \code{nlag}, or list with elements \code{msd}, \code{alpha}, \code{D}.
}
\description{
Sample estimate of mean squared displacement.
}
