% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msd_fit.R
\name{msd_fit}
\alias{msd_fit}
\title{Sample estimate of mean squared displacement.}
\usage{
msd_fit(Xt, nlag, demean = TRUE)
}
\arguments{
\item{Xt}{Vector or matrix of trajectory positions (each column is a different coordinate).  The trajectory is assumed to be sampled at a constant frequency.}

\item{nlag}{Number of MSD lags to calculate.}

\item{demean}{Logical; whether or not to remove the mean of \code{Xt} as estimated by linear drift.}
}
\value{
Sample MSD vector of length \code{nlag}.
}
\description{
Sample estimate of mean squared displacement.
}
