% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msd_fit.R
\name{msd_fit}
\alias{msd_fit}
\title{Sample estimate of mean squared displacement.}
\usage{
msd_fit(Xt, nlag, demean = TRUE)
}
\arguments{
\item{Xt}{Matrix of trajectory positions, where each row is an observation and each column is a measurement coordinate.  The trajectory is assumed to be sampled at a constant frequency.}

\item{nlag}{Number of MSD lags to calculate.}

\item{demean}{Logical; whether or not to subtract from \code{Xt} a regression-based estimate of its linear drift.}
}
\value{
Sample MSD vector of length \code{nlag}.
}
\description{
Sample estimate of mean squared displacement.
}
\examples{
# generates a 2-d series of fBM.
alpha <- 0.8
dt <- 1/60
N <- 1800
acf1 <- fbm_acf(alpha, dt, N)
dX <- rSnorm(2, acf = acf1)
Xt <- apply(dX, 2, cumsum)
# Compute the MSD of Xt
nlag <- 600
msd1 <- msd_fit(Xt, nlag = nlag)
}
