% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msd_ls.R
\name{msd_ls}
\alias{msd_ls}
\title{Least-squares power-law fit to sample mean square displacement curves.}
\usage{
msd_ls(msd, dT, pooled = TRUE, logw = TRUE)
}
\arguments{
\item{msd}{Vector or matrix of sample MSDs, each column corresponding to a different trajectory.}

\item{dT}{Scalar time increment between MSD observations (see Details).}

\item{pooled}{Logical; whether to calculate separate regressions for each MSD or pool all of them together.}

\item{logw}{Logical; whether or not regression should be log-weighted (see Details).}
}
\value{
A vector of length 2 if \code{pooled = TRUE}, or a 2-row matrix of \code{alpha} and \code{D} values if \code{pooled = FALSE}.
}
\description{
Least-squares power-law fit to sample mean square displacement curves.
}
\details{
A power-law fit to the empirical MSD is calculated by performing the linear regression
\preformatted{
log(msd) ~ log(D) + alpha * log(time).
}
From a curve fitting perspective, we seek to estimate the "best" power-law describing the MSD, i.e., if MSD(t) is the true MSD at time t, then the best power law minimizes
\preformatted{
integral_0^Inf | log MSD(t) - log(D) - alpha * log(t) |^2 w(t) dt,
}
where \code{w(t) > 0} is a weight function.  The default value of \code{logw = FALSE} uses \code{w(t) = 1}, but this doesn't give the best MSD fit on the log-log scale, as there are exponentially more points as we move right in the graph, such that the right side of the graph will dominate the fit.  Setting \code{logw = TRUE} uses uniform weighting on the log-scale, which corresponds to \code{w(t) = 1/t}.

While the MSDs provided to the \code{msd} argument must all share a common time increment \code{dT}, they can have different lengths, in which case missing time points in the \code{msd} matrix should be indicated by \code{NA}.
}
