% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prony-acf.R
\name{prony.acf}
\alias{prony.acf}
\title{Prony Acf.}
\usage{
prony.acf(lambda, N, dt, ...)
}
\arguments{
\item{lambda}{coefficients of the sum of OU processes.}

\item{N}{number of samples.}

\item{dt}{interobservation time.}

\item{...}{Additional arguments to \code{prony.coeff} for obtaining the BM-OU coefficients.}
}
\value{
Vector of autocorrelations.
}
\description{
GLE with 0 mass and 0 external potential.
}
\details{
process \code{x} is defined as
\code{vsigma * int_{-infty}^t gamma(t-s)  x_s d s = k_B T vsigma * F_t}
where the force noise F_t is a sum of independent OU processes, ie,
cov(F_s, F_{t+s}) = gamma(t) = exp(-alpha1 t) + ... + exp(-alphaK t).
the resulting process is
Y_t = C1 X_1t + ... + CK X_Kt,
where X_1t = B_1t and X_it, i > 1 are OU processes,
d X_it = -ri X_it dt + d B_it,
and all B_it are independent.
note that cov(X_i0, X_it) = (2*ri)^{-1} * exp(-ri * t).
}
