% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prony-msd.R
\name{prony.msd}
\alias{prony.msd}
\title{Prony MSD}
\usage{
prony.msd(t, alpha, nSteps = 100, tol = 0, r, C)
}
\arguments{
\item{t}{times at which to calculate the MSD.}

\item{alpha}{coefficients of the sum of OU processes.}

\item{nSteps}{number of steps in mode finding (defaults to 100)}

\item{tol}{tolerance for mode finding (defaults to 0, i.e. always take nSteps)}

\item{r}{roots that can accelerate computations.}

\item{C}{coefficients that can accelerate computations, can't pass C without r.}
}
\value{
list containing msd, C, r and lmsdp
}
\description{
MSD of a GLE with 0 mass and 0 external potential.
}
\details{
$$ vsigma * \int_{-\infty}^t \gamma(t-s) \dot x_s \ud s = k_B T vsigma * F_t
where the force noise F_t is a sum of independent OU processes, ie,
cov(F_s, F_{t+s}) = gamma(t) = exp(-alpha1 t) + ... + exp(-alphaK t).
the resulting process is
Y_t = C1 X_1t + ... + CK X_Kt,
where X_1t = B_1t and X_it, i > 1 are OU processes,
d X_it = -ri X_it dt + d B_it,
and all B_it are independent.
note that cov(X_i0, X_it) = (2*ri)^{-1} * exp(-ri * t)
}
