% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subdiff-package.R
\docType{package}
\name{subdiff-package}
\alias{subdiff}
\alias{subdiff-package}
\title{Inference for Subdiffusive Particle Tracking.}
\description{
Tools for implementing various models for particle subdiffusion in biological fluids.  In addition to the well-known semiparametric least-squares estimator based on the mean square displacement (MSD), the package provides functions for simulation, inference, and goodness-of-fit for two fully-parametric subdiffusion models: fractional Brownian motion and the generalized Langevin equation (GLE) model with Rouse memory kernel.  A generic model class allows users to easily implement custom subdiffusion models, with a ready-made framework for drift, high-frequency error correction, and efficient maximum likelihood estimation.
}
\author{
\strong{Maintainer}: Martin Lysy \email{mlysy@uwaterloo.ca}

Authors:
\itemize{
  \item Yun Ling
}

}
