% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_Sigma.R
\name{trans_Sigma}
\alias{trans_Sigma}
\alias{itrans_Sigma}
\title{Unconstraining transformation for variance matrices.}
\usage{
trans_Sigma(Sigma)

itrans_Sigma(lambda)
}
\arguments{
\item{Sigma}{Variance matrix on the regular scale.}

\item{lambda}{Variance matrix on the normalized scale (see 'Details').}
}
\value{
Variance matrix on the regular or unconstrained scale (see 'Details').
}
\description{
Unconstraining transformation for variance matrices.
}
\details{
The unconstraining transformation of a variance matrix is the so-called log-Cholesky decomposition.  Namely, the log-Cholesky decomposition of a variance matrix \code{Sigma} is a vector \code{lambda} corresponding to the upper triangular Cholesky factor, of which we take the log of the diagonal and then concatenate the non-zero elements in column-major order.  The exact calculation is given by:

\if{html}{\out{<div class="sourceCode">}}\preformatted{lambda <- chol(Sigma)
diag(lambda) <- log(diag(lambda))
lambda <- lambda[upper.tri(lambda,diag=TRUE)]
}\if{html}{\out{</div>}}

The function \code{trans_Sigma()} converts \code{Sigma} to \code{lambda}, whereas \code{itrans_Sigma()} performs the inverse transformation from \code{lambda} to \code{Sigma}.
}
